\name{2019_correlationVecchia-package}
\alias{2019_correlationVecchia-package}
\alias{2019_correlationVecchia}
\docType{package}
\title{
\packageTitle{2019_correlationVecchia}
}
\description{
\packageDescription{2019_correlationVecchia}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{2019_correlationVecchia}
\packageIndices{2019_correlationVecchia}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{2019_correlationVecchia}

Maintainer: \packageMaintainer{2019_correlationVecchia}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
