% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{baseline_1_multivariate_specify}
\alias{baseline_1_multivariate_specify}
\title{The first baseline approximation for multivaraite cases}
\usage{
baseline_1_multivariate_specify(locs, m)
}
\arguments{
\item{locs}{A matrix of locations}

\item{m}{A size of conditioning sets}
}
\value{
list
}
\description{
The first baseline approximation for multivaraite cases
}
\examples{
locs1  <- matrix(runif(20), 10, 2)
locs2  <- matrix(runif(30), 15, 2)
locs   <- list(locs1 = locs1, locs2 = locs2)
m      <- 5

# baseline 1 = separately E-maxmin ordering + half-and-half E-NN conditioning
# baseline 2 = separately E-maxmin ordering + unified E-NN conditioning 
# baseline 3 = separately E-maxmin ordering + separately E-NN conditioning
# baseline 4 = separately E-maxmin ordering + unified C-NN conditioning

# Note 1 = For all baseline approximations, the first process is not relevant to the second one
# Note 2 = For the third baseline approximations, two processes are totally irrelevant

### ordering
ord1 <- baseline_1_multivariate_specify(locs = locs, m = m)$ord
ord2 <- baseline_2_multivariate_specify(locs = locs, m = m)$ord
ord3 <- baseline_3_multivariate_specify(locs = locs, m = m)$ord
ord4 <- baseline_4_multivariate_specify(locs = locs, m = m, 
                                        covmodel = cov_bivariate_flexMatern, 
                                        sigma.mat = matrix(c(1, 0.5, 0.5, 1), 2, 2), 
                                        nu.mat = matrix(0.5, 2, 2), 
                                        alpha.mat = matrix(1, 2, 2))$ord

all.equal(ord1, ord2) ; all.equal(ord1, ord3) ; all.equal(ord1, ord4)

### conditioning
ind <- seq(from = m, to = 1, by = -1)
cond1 <- baseline_1_multivariate_specify(locs = locs, m = m)$U.prep$revNNarray[, ind]
cond2 <- baseline_2_multivariate_specify(locs = locs, m = m)$U.prep$revNNarray[, ind]
cond3 <- baseline_3_multivariate_specify(locs = locs, m = m)$U.prep$revNNarray[, ind]
cond4 <- baseline_4_multivariate_specify(locs = locs, m = m, 
                                         covmodel = cov_bivariate_flexMatern, 
                                         sigma.mat = matrix(c(1, 0.5, 0.5, 1), 2, 2), 
                                         nu.mat = matrix(0.5, 2, 2), 
                                         alpha.mat = matrix(1, 2, 2))$U.prep$revNNarray[, ind]

cond1 == cond2 # similar
cond1 == cond3 # different
cond3 == cond4 # similar
}
