% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{baseline_1_spacetime_specify}
\alias{baseline_1_spacetime_specify}
\title{The first baseline approximation for spatio-temporal GPs (T-ord + T-NN)}
\usage{
baseline_1_spacetime_specify(locs, m, coordinate = NULL)
}
\arguments{
\item{locs}{A matrix of spatio-temporal locations}

\item{m}{A size of conditioning sets}

\item{coordinate}{integer or vector of integers in \code{1,...,d}}
}
\value{
vecchia.approx
}
\description{
The first baseline approximation for spatio-temporal GPs (T-ord + T-NN)
}
\examples{
locs      <- matrix(runif(30), nrow = 10, ncol = 3)
m         <- 4
covparms  <- c(1, 0.1, 1.0)

# baseline 1 = time-based ordering + the m previous obs conditioning (T-ord + T-NN)
# baseline 2 = time-based ordering + E-NN 
#              without distinction between time and space coordinates (T-ord + E-NN)
# baseline 3 = time-based ordering + C-NN conditioning (T-ord + C-NN)

### ordering
ord1 <- baseline_1_spacetime_specify(locs = locs, m = m)$ord
ord2 <- baseline_2_spacetime_specify(locs = locs, m = m)$ord
ord3 <- baseline_3_spacetime_specify(locs = locs, m = m,
                                     covmodel = cov_expo_spacetime,
                                     covparms = covparms)$ord

all.equal(ord1, ord2) ; all.equal(ord1, ord3)

### conditioning
ind <- seq(from = m, to = 1, by = -1)
cond1 <- baseline_1_spacetime_specify(locs = locs, m = m)$U.prep$revNNarray[, ind]
cond2 <- baseline_2_spacetime_specify(locs = locs, m = m)$U.prep$revNNarray[, ind]
cond3 <- baseline_3_spacetime_specify(locs = locs, m = m,
                                      covmodel = cov_expo_spacetime,
                                      covparms = covparms)$U.prep$revNNarray[, ind]

cond1
cond2
cond3
}
