% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{baseline_3_spacetime_specify}
\alias{baseline_3_spacetime_specify}
\title{The third baseline approximation for spatio-temporal cases}
\usage{
baseline_3_spacetime_specify(locs, m, covmodel, covparms)
}
\arguments{
\item{locs}{A matrix of spatio-temporal locations}

\item{m}{A size of conditioning sets}

\item{covmodel}{covariance function}

\item{covparms}{A numeric vector of covariance parameters}
}
\value{
list
}
\description{
The third baseline approximation for spatio-temporal cases
}
\examples{
locs <- matrix(runif(30), nrow = 10, ncol = 3)
m <- 4
covparms <- c(1, 0.75, 50, 25)

# baseline 1 = time-based ordering + the m previous obs conditioning
# baseline 2 = time-based ordering + E-NN without 
#                                      distinction between time and space coordinates
# baseline 3 = time-based ordering + C-NN conditioning

### ordering
ord1 <- baseline_1_spacetime_specify(locs = locs, m = m)$ord
ord2 <- baseline_2_spacetime_specify(locs = locs, m = m)$ord
ord3 <- baseline_3_spacetime_specify(locs = locs, m = m, 
                                     covmodel = cov_spacetime_expo, 
                                     covparms = covparms)$ord

all.equal(ord1, ord2) ; all.equal(ord1, ord3)

### conditioning
ind <- seq(from = m, to = 1, by = -1)
cond1 <- baseline_1_spacetime_specify(locs = locs, m = m)$U.prep$revNNarray[, ind]
cond2 <- baseline_2_spacetime_specify(locs = locs, m = m)$U.prep$revNNarray[, ind]
cond3 <- baseline_3_spacetime_specify(locs = locs, m = m, 
                                      covmodel = cov_spacetime_expo, 
                                      covparms = covparms)$U.prep$revNNarray[, ind]

cond1
cond2 
cond3
}
