% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{baseline_mulv_specify}
\alias{baseline_mulv_specify}
\title{Baseline approximation for multivariate GPs}
\usage{
baseline_mulv_specify(approx, locs, m, ...)
}
\arguments{
\item{approx}{A type of approximation = 1 (S-E-MM + HH-E-NN), 2 (S-E-MM + J-E-NN), 3 (S-E-MM + S-E-NN), or 4 (S-E-MM + C-NN)}

\item{locs}{A matrix of locations}

\item{m}{A size of conditioning sets}

\item{...}{Ellipsis}
}
\value{
vecchia.approx
}
\description{
Baseline approximation for multivariate GPs
}
\examples{
locs1  <- matrix(runif(20), 10, 2)
locs2  <- matrix(runif(30), 15, 2)
locs   <- list(locs1 = locs1, locs2 = locs2)
m      <- 5

ord1 <- baseline_mulv_specify(approx = 1, locs = locs, m = m)$ord
ord2 <- baseline_mulv_specify(approx = 2, locs = locs, m = m)$ord
ord3 <- baseline_mulv_specify(approx = 3, locs = locs, m = m)$ord
ord4 <- baseline_mulv_specify(approx = 4, locs = locs, m = m,
                              covmodel = cov_flexMatern_biv,
                              sigma.mat = matrix(c(1, 0.5, 0.5, 1), 2, 2),
                              nu.mat = matrix(0.5, 2, 2),
                              alpha.mat = matrix(1, 2, 2))$ord

all.equal(ord1, ord2) ; all.equal(ord1, ord3) ; all.equal(ord1, ord4)
}
