% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{baseline_spacetime_specify}
\alias{baseline_spacetime_specify}
\title{Baseline approximation for spatio-temporal GPs}
\usage{
baseline_spacetime_specify(approx, locs, m, coordinate = NULL, ...)
}
\arguments{
\item{approx}{A type of approximation = 1 (T-ord + T-NN), 2 (T-ord + E-NN), or 3 (T-ord + C-NN)}

\item{locs}{A matrix of spatio-temporal locations}

\item{m}{A size of conditioning sets}

\item{coordinate}{integer or vector of integers in \code{1,...,d}}

\item{...}{Ellipsis}
}
\value{
vecchia.approx
}
\description{
Baseline approximation for spatio-temporal GPs
}
\examples{
locs      <- matrix(runif(30), nrow = 10, ncol = 3)
m         <- 4
covparms  <- c(1, 0.1, 1.0)

### ordering
ord1 <- baseline_spacetime_specify(approx = 1, locs = locs, m = m)$ord
ord2 <- baseline_spacetime_specify(approx = 2, locs = locs, m = m)$ord
ord3 <- baseline_spacetime_specify(approx = 3, locs = locs, m = m,
                                   covmodel = cov_expo_spacetime,
                                   covparms = covparms)$ord

all.equal(ord1, ord2) ; all.equal(ord1, ord3)
}
