% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cal_matern_ns}
\alias{cal_matern_ns}
\title{For fast calculation of nonstationary Matern covariance (See cov_matern_ns)}
\usage{
cal_matern_ns(locs, sigvec, nuvec, kerList)
}
\arguments{
\item{locs}{A numeric matrix of locations}

\item{sigvec}{A vector of variance parameters}

\item{nuvec}{A vector of smoothness parameters}

\item{kerList}{A list of kernel matrices}
}
\value{
The nonstationary Matern covariance matrix of the locations provided
}
\description{
For fast calculation of nonstationary Matern covariance (See cov_matern_ns)
}
