% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{conditioning_b1_sptm_Rcpp}
\alias{conditioning_b1_sptm_Rcpp}
\title{Nearest-Neighbor (NN) conditioning for baseline_1_spacetime_specify()}
\usage{
conditioning_b1_sptm_Rcpp(n, m)
}
\arguments{
\item{n}{Number of locations}

\item{m}{Number of nearby points to condition on}
}
\value{
A matrix of indices giving NN conditioning sets
}
\description{
Nearest-Neighbor (NN) conditioning for baseline_1_spacetime_specify()
}
