% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{conditioning_m_Rcpp}
\alias{conditioning_m_Rcpp}
\title{Nearest-Neighbor (NN) conditioning with respect to user-specified distance matrix}
\usage{
conditioning_m_Rcpp(m, d)
}
\arguments{
\item{m}{Number of nearby points to condition on}

\item{d}{A matrix of distances between locations (for instance, 1 - cor)}
}
\value{
A matrix of indices giving NN conditioning sets
}
\description{
Nearest-Neighbor (NN) conditioning with respect to user-specified distance matrix
}
