% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia.R
\name{conditioning_nn}
\alias{conditioning_nn}
\title{Nearest-Neighbor (NN) conditioning with respect to user-specified distance matrix}
\usage{
conditioning_nn(m, d)
}
\arguments{
\item{m}{Number of nearby points to condition on}

\item{d}{A matrix of distances between locations (for instance, 1-rho)}
}
\value{
A matrix of indices giving NN conditioning sets
}
\description{
Nearest-Neighbor (NN) conditioning with respect to user-specified distance matrix
}
\examples{
locs <- matrix(runif(100 * 2), 100, 2)
d <- as.matrix(dist(locs))
which(conditioning_nn(5, d) != GpGp::find_ordered_nn_brute(locs, 5), arr.ind = TRUE)
which(conditioning_nn(5, d) != GpGp::find_ordered_nn(locs, 5), arr.ind = TRUE)

n <- 10 ; m <- 3

locs <- matrix(runif(n * 2), n, 2)
d <- as.matrix(dist(locs))

conditioning_nn(m, d)
conditioning_nn_Rcpp(m, d) + 1

require(microbenchmark)
microbenchmark(conditioning_nn(m, d), conditioning_nn_Rcpp(m, d), times = 2)
}
