% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{cormat_hnm}
\alias{cormat_hnm}
\title{Correlation matrix of Hierarchical Normal Model (HNM)}
\usage{
cormat_hnm(output)
}
\arguments{
\item{output}{output of \code{generate_hnm()}}
}
\value{
Correlation matrix
}
\description{
Correlation matrix of Hierarchical Normal Model (HNM)
}
\examples{
output <- generate_hnm(mu0 = 0, sig20 = 1, k = 3, mlt = 2)

length(output$y)

covmat <- covmat_hnm(output)
cormat <- cormat_hnm(output)

fields::imagePlot(cormat)
}
