% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmodels.R
\name{corr_derivative_matern_4.5_2d}
\alias{corr_derivative_matern_4.5_2d}
\title{Matern correlation matrix of (GP, dGP/dx1, dGP/dx2) with respect to the smoothness parameter of 4.5 in 2-dimensional domain}
\usage{
corr_derivative_matern_4.5_2d(locs, covparms, tol = .Machine$double.eps)
}
\arguments{
\item{locs}{A matrix of locations}

\item{covparms}{A numerical vector with covariance parameters = c(sigma2, range)}

\item{tol}{A numerical tolerance for the damped sine function. At \code{.Machine$double.eps} by default}
}
\value{
A matrix with \code{3*n} rows and \code{3*n} columns
}
\description{
Matern correlation matrix of (GP, dGP/dx1, dGP/dx2) with respect to the smoothness parameter of 4.5 in 2-dimensional domain
}
\examples{
locs <- matrix(runif(8), 4, 2)
covparms <- c(1, 0.1)
tol <- .Machine$double.eps

M <- corr_derivative_matern_4.5_2d(locs = locs, covparms = covparms, tol = tol)

t(M) == M # M is symmetric
eigen(M)$values # M is positive definite
}
