% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmodels.R
\name{corr_derivative_squared_expo_1d}
\alias{corr_derivative_squared_expo_1d}
\title{Squared exponential correlation matrix of (GP, dGP/dx) in 1-dimensional domain}
\usage{
corr_derivative_squared_expo_1d(locs, covparms, tol = .Machine$double.eps)
}
\arguments{
\item{locs}{A matrix of locations}

\item{covparms}{A numerical vector with covariance parameters = c(sigma2, range)}

\item{tol}{A numerical tolerance for the damped sine function. At \code{.Machine$double.eps} by default}
}
\value{
A matrix with \code{2*n} rows and \code{2*n} columns
}
\description{
Squared exponential correlation matrix of (GP, dGP/dx) in 1-dimensional domain
}
\examples{
locs <- matrix(runif(4), 4, 1)
covparms <- c(1, 0.1)
tol <- .Machine$double.eps

M <- corr_derivative_squared_expo_1d(locs, covparms, tol)

t(M) == M # M is symmetric
eigen(M)$values # M is positive definite
}
