\name{correlationVecchia-package}
\alias{correlationVecchia-package}
\alias{correlationVecchia}
\docType{package}
\title{
\packageTitle{correlationVecchia}
}
\description{
\packageDescription{correlationVecchia}
}
\details{
This is a private R package for research.
% The DESCRIPTION file:
% \packageDESCRIPTION{vecchia}
% \packageIndices{vecchia}
% ~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Myeongjong Kang and Matthias Katzfuss\cr Maintainer: NA
}
\references{
This is a private R package for research.
}
% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
This is a private R package for research.
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
1 + 1
% ~~ simple examples of the most important functions ~~
}
