% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmodels.R
\name{cov_derivative_matern_2.5_2d}
\alias{cov_derivative_matern_2.5_2d}
\title{Matern covariance matrix of (GP, dGP/dx, dGP/dy) with respect to the smoothness parameter of 2.5 in 2 - dimensional domain}
\usage{
cov_derivative_matern_2.5_2d(locs, covparms, tol = .Machine$double.eps)
}
\arguments{
\item{locs}{A matrix of locations}

\item{covparms}{A numerical vector with covariance parameters = c(sigma2, range)}

\item{tol}{A numerical tolerance for the damped sine function. At \code{.Machine$double.eps} by default}
}
\value{
A matrix with \code{3*n} rows and \code{3*n} columns
}
\description{
Matern covariance matrix of (GP, dGP/dx, dGP/dy) with respect to the smoothness parameter of 2.5 in 2 - dimensional domain
}
\examples{
locs <- matrix(runif(6), 3, 2)
covparms <- c(1, 1)
coord <- 1
tol <- .Machine$double.eps

covmat <- cov_derivative_matern_2.5_2d(locs = locs, covparms = covparms, tol = tol)

eigen(covmat)$values
}
