% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cov_expo_aniso_cpp}
\alias{cov_expo_aniso_cpp}
\title{Anisotropic exponential covariance function (C++ version)}
\usage{
cov_expo_aniso_cpp(x1, x2, covparms)
}
\arguments{
\item{x1}{First location (vector)}

\item{x2}{Second location (vector)}

\item{covparms}{A numerical vector with covariance parameters in the form (variance, range, anisotropy for the first dimension)}
}
\value{
The anisotropic exponenital covariance between the locations provided
}
\description{
Anisotropic exponential covariance function (C++ version)
}
