% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmodels.R
\name{cov_expo_spacetime}
\alias{cov_expo_spacetime}
\title{Space-time exponential covariance function}
\usage{
cov_expo_spacetime(locs, covparms)
}
\arguments{
\item{locs}{A numerical matrix with \code{n} rows and \code{d} columns. Each row of locs gives a point in R^d}

\item{covparms}{A numerical vector with covariance parameters = (sigma2, spatial range, temporal range)}
}
\value{
A matrix with \code{n} rows and \code{n} columns, with the \code{(i, j)} entry containing the space-time exponenital covariance between observations \code{locs[i, ]} and \code{locs[j, ]}
}
\description{
From a location matrix \code{locs} and a vector with covariance parameters \code{covparms}, this function returns a space-time exponential covariance matrix.
}
\examples{
set.seed(5152021)

locs <- matrix(runif(30), 10, 3)
locs

covmat <- cov_expo_spacetime(locs = locs, covparms = c(1, 0.1, 1.0))
fields::image.plot(covmat)

covmat <- cov_expo_spacetime(locs = locs, covparms = c(1, 0.1, 1.0))
fields::image.plot(covmat)
}
