% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cov_expo_spacetime_cpp}
\alias{cov_expo_spacetime_cpp}
\title{Space-time exponential covariance function (C++ version)}
\usage{
cov_expo_spacetime_cpp(x1, x2, covparms)
}
\arguments{
\item{x1}{First location (vector)}

\item{x2}{Second location (vector)}

\item{covparms}{A numerical vector with covariance parameters in the form (variance, spatial range, temporal range)}
}
\value{
The space-time exponenital covariance between the locations provided
}
\description{
Space-time exponential covariance function (C++ version)
}
