% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmodels.R
\name{cov_GK_mulv}
\alias{cov_GK_mulv}
\title{Flexible Matern covariance function for Multivariate Processes (Genton and Kleiber, 2015)}
\usage{
cov_GK_mulv(
  locs,
  sigma.mat = NULL,
  p = NULL,
  rho = NULL,
  nu.mat = NULL,
  nu.vec = NULL,
  alpha.mat = NULL,
  alpha = NULL
)
}
\arguments{
\item{locs}{A matrix of locations}

\item{sigma.mat}{A numerical \code{p} by \code{p} matrix of collocated covariance coefficients at \code{NULL} by default}

\item{p}{A number of processes at \code{NULL} by default. Please use this argument with the next argument rho}

\item{rho}{A coefficient for defining a matrix of collocated covariance coefficients as an autoregressive covariance matrix. Please use this argument with the previous argument p}

\item{nu.mat}{A numerical \code{p} by \code{p} matrix of smoothness parameters at \code{NULL} by default}

\item{nu.vec}{A numeric vector at \code{NULL} by default. If nu.vec is used instead of nu.mat, the \code{(i, j)}-th entry a matrix of smoothness parametes is defined by \code{(nu_i + nu_j)/2}}

\item{alpha.mat}{A numerical \code{p} by \code{p} matrix of scale parameters}

\item{alpha}{Numeric at NULL by default. If alpha is used instead of alpha.mat, a matrix of scale parameters is defined by a constant matrix of alpha}
}
\value{
A flexible Matern covariance matrix
}
\description{
Flexible Matern covariance function for Multivariate Processes (Genton and Kleiber, 2015)
}
\examples{
n <- 3 ; d <- 2 ; p <- 2
locs <- matrix(runif(n * d), n, d)
sigma.mat <- diag(p) ; nu.mat <- matrix(0.5, p, p) ; alpha.mat <- matrix(1, p, p)

covmat <- cov_flexMatern_mulv(locs = locs,
                              sigma.mat = sigma.mat,
                              nu.mat = nu.mat,
                              alpha.mat = alpha.mat)
covmat_each <- cov_GK_mulv(locs = locs,
                           sigma.mat = sigma.mat,
                           nu.mat = nu.mat,
                           alpha.mat = alpha.mat)
covmat_each_bruteforce <- cov_GK_mulv_bruteforce(locs = locs,
                                                 sigma.mat = sigma.mat,
                                                 nu.mat = nu.mat,
                                                 alpha.mat = alpha.mat)

covmat_matched <- matrix(NA, n * p, n * p)
for(i in 1:p){
  for(j in 1:p){
    covmat_matched[seq(from = 1 + n * (i - 1), to = n * i, by = 1),
                   seq(from = 1 + n * (j - 1), to = n * j, by = 1)] <-
                   covmat_each[seq(from = i, to = i + p * (n - 1), by = p),
                               seq(from = j, to = j + p * (n - 1), by = p)]
  }
}

covmat
covmat_each
covmat_each_bruteforce

identical(covmat_each, covmat_each_bruteforce)
identical(covmat, covmat_matched)
}
\references{
Genton, Marc G., and William Kleiber. "Cross-covariance functions for multivariate geostatistics." Statistical Science (2015): 147-163.
}
