% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cov_latentDim_triv_cpp}
\alias{cov_latentDim_triv_cpp}
\title{Isotropic exponential covariance function for trivariate Gaussian Processes using latent dimensions (C++ version)}
\usage{
cov_latentDim_triv_cpp(x1, x2, covparms)
}
\arguments{
\item{x1}{First location (vector)}

\item{x2}{Second location (vector)}

\item{covparms}{A numerical vector with covariance parameters = (sigma2, range, latent dimension, x1's id, x2's id). ID can be 1 or 2, or 3.}
}
\value{
The isotropic exponential covariance between the locations provided with respect to latent dimension and their ids
}
\description{
Isotropic exponential covariance function for trivariate Gaussian Processes using latent dimensions (C++ version)
}
