% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmodels.R
\name{cov_matern}
\alias{cov_matern}
\title{Matern covariance function with smoothness parameter of 2.5}
\usage{
cov_matern(locs, covparms, tol = .Machine$double.eps)
}
\arguments{
\item{locs}{A matrix of locations}

\item{covparms}{A numerical vector with covariance parameters = (sigma2, range)}

\item{tol}{A numerical tolerance for the damped sine function. At \code{.Machine$double.eps} by default}
}
\value{
A matrix with \code{n} rows and \code{n} columns, with the \code{(i, j)} entry containing the matern covariance between observations \code{locs[i, ]} and \code{locs[j, ]}
}
\description{
Matern covariance function with smoothness parameter of 2.5
}
\examples{
cov_matern(locs = matrix(c(1, 0), 2, 1), covparms = c(1, 1, 2.5))[2, 1]
cov_matern_2.5(locs = matrix(c(1, 0), 2, 1), covparms = c(1, 1))[2, 1]
fields::Matern(d = 1, range = 1, nu = 2.5)
fields::Matern(d = sqrt(5), range = 1, nu = 2.5)
}
