% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmodels.R
\name{cov_matern_3.5}
\alias{cov_matern_3.5}
\title{Matern covariance function with smoothness parameter of 3.5}
\usage{
cov_matern_3.5(locs, covparms)
}
\arguments{
\item{locs}{A matrix of locations}

\item{covparms}{A numerical vector with covariance parameters = (sigma2, range)}
}
\value{
A matrix with \code{n} rows and \code{n} columns, with the \code{(i, j)} entry containing the matern covariance with smoothness parameter of 2.5 between observations \code{locs[i, ]} and \code{locs[j, ]}
}
\description{
Matern covariance function with smoothness parameter of 3.5
}
\examples{
locs <- matrix(runif(10), 5, 2)
fields::Matern(d = fields::rdist(locs), range = 1, smoothness = 3.5)
cov_matern_3.5(locs, covparms = c(1, 1))

locs <- matrix(runif(10), 5, 2)
fields::Matern(d = fields::rdist(locs), range = 1, smoothness = 3.5)
cov_matern_3.5(locs / sqrt(7), covparms = c(1, 1))
}
