% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmodels.R
\name{cov_matern_aniso}
\alias{cov_matern_aniso}
\title{Anisotropic Matern covariance function}
\usage{
cov_matern_aniso(locs, covparms, tol = .Machine$double.eps)
}
\arguments{
\item{locs}{A matrix of locations}

\item{covparms}{A numerical vector with covariance parameters = (sigma2, range, smoothness, anisotropy)}

\item{tol}{A numerical tolerance for the damped sine function. At \code{.Machine$double.eps} by default}
}
\value{
A matrix with \code{n} rows and \code{n} columns, with the \code{(i, j)} entry containing the anisotropic matern covariance between observations \code{locs[i, ]} and \code{locs[j, ]}
}
\description{
Anisotropic Matern covariance function
}
\examples{
locs <- matrix(runif(4), 2, 2)

cov_matern_aniso(locs, covparms = c(1, 1, 0.5, 5))
cov_matern_aniso_cpp(locs[1, ], locs[2, ], covparms = c(1, 1, 0.5, 5))
}
