% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cov_matern_iso_cpp}
\alias{cov_matern_iso_cpp}
\title{Isotropic Matern covariance function (C++ version)}
\usage{
cov_matern_iso_cpp(x1, x2, covparms)
}
\arguments{
\item{x1}{First location (vector)}

\item{x2}{Second location (vector)}

\item{covparms}{A numerical vector with covariance parameters in the form (sigma2, range, smoothness)}
}
\value{
The isotropic Matern covariance between the locations provided
}
\description{
Isotropic Matern covariance function (C++ version)
}
