% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmodels.R
\name{cov_matern_ns}
\alias{cov_matern_ns}
\title{Nonstationary Matern covariance function}
\usage{
cov_matern_ns(locs, sigma, smoothness, kernel)
}
\arguments{
\item{locs}{A numerical matrix with \code{n} rows and \code{p} columns. Each row of locs1 gives a point in R^d}

\item{sigma}{A numeric function for spatially varying standard deviation \eqn{\sigma( loc )}}

\item{smoothness}{A numeric function for spatially varying smoothness \eqn{\nu( loc )}}

\item{kernel}{A matrix-valued function for spatially varying (local) geometric anisotropy \eqn{\Sigma( loc )}}
}
\value{
A matrix with \code{n} rows and \code{n} columns, with the \code{(i, j)} entry containing the nonstationary Matern covariance between observations \code{locs[i, ]} and \code{locs[j, ]}
}
\description{
From a location matrix \code{locs} and functions \code{sigma}, \code{smoothness}, \code{kernel}, this function returns a nonstationary matern covariance matrix.
}
\examples{
aniso <- function(loc) {

  d         <- length(loc)

  a         <- function(loc) 1
  b         <- function(loc) 10
  angle     <- function(loc) pi/2

  eta       <- angle(loc)
  rot.mat   <- matrix(c(cos(eta), sin(eta), -sin(eta), cos(eta)), nrow = d, ncol = d, byrow = TRUE)

  range     <- c(a(loc)^(-2), b(loc)^(-2))

  return( t(rot.mat) \%*\% diag(range, nrow = d) \%*\% rot.mat )
}

sigma <- function(loc) determinant(aniso(loc), logarithm = FALSE)[[1]][1]^0.25

smoothness <- function(loc) 0.5

locs <- matrix(runif(6), 3, 2)

cov_matern_ns(locs = locs,
              sigma = sigma, smoothness = smoothness, kernel = aniso)

cov_matern_ns_bruteforce(locs1 = locs, locs2 = locs,
                        sigma = sigma, smoothness = smoothness, kernel = aniso)

\dontrun{

  locs <- matrix(runif(800), 400, 2)

  microbenchmark::microbenchmark(
    cov_matern_ns(locs = locs,
                  sigma = sigma, smoothness = smoothness, kernel = aniso),
    cov_matern_ns_bruteforce(locs1 = locs, locs2 = NULL,
                             sigma = sigma, smoothness = smoothness, kernel = aniso),
    times = 5)

}
}
\references{
\itemize{
\item (Main) Katzfuss, Matthias. "Bayesian nonstationary spatial modeling for very large datasets." Environmetrics 24.3 (2013): 189-200.
\item Paciorek, Christopher J., and Mark J. Schervish. "Spatial modelling using a new class of nonstationary covariance functions." Environmetrics: The official journal of the International Environmetrics Society 17.5 (2006): 483-506.
\item Stein, Michael L. "Nonstationary spatial covariance functions." Unpublished technical report (2005).
}
}
