% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cov_matern_ns_cpp}
\alias{cov_matern_ns_cpp}
\title{Nonstationary Matern covariance function (C++ version)}
\usage{
cov_matern_ns_cpp(x1, x2, sigma, smoothness, kernel)
}
\arguments{
\item{x1}{First location (vector)}

\item{x2}{Second location (vector)}

\item{sigma}{A numeric function for spatially varying standard deviation \eqn{\sigma( loc )}}

\item{smoothness}{A numeric function for spatially varying smoothness \eqn{\nu( loc )}}

\item{kernel}{A matrix-valued function for spatially varying (local) geometric anisotropy \eqn{\Sigma( loc )}}
}
\value{
The nonstationary Matern covariance between the locations provided
}
\description{
Nonstationary Matern covariance function (C++ version)
}
