% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cov_matern_ns_forR}
\alias{cov_matern_ns_forR}
\title{Nonstationary Matern covariance function (similar to R version, but slower than R version)}
\usage{
cov_matern_ns_forR(locs1, locs2, sigma, smoothness, kernel)
}
\arguments{
\item{locs1}{A numerical matrix with \code{n} rows and \code{p} columns. Each row of locs1 gives a point of the first set in R^d}

\item{locs2}{A numerical matrix with \code{n} rows and \code{p} columns. Each row of locs2 gives a point of the second set in R^d}

\item{sigma}{A numeric function for spatially varying standard deviation \eqn{\sigma( loc )}}

\item{smoothness}{A numeric function for spatially varying smoothness \eqn{\nu( loc )}}

\item{kernel}{A matrix-valued function for spatially varying (local) geometric anisotropy \eqn{\Sigma( loc )}}
}
\value{
The nonstationary Matern covariance between the locations provided
}
\description{
Nonstationary Matern covariance function (similar to R version, but slower than R version)
}
