% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{cov_matern_scaledim}
\alias{cov_matern_scaledim}
\title{Wrapper function for Matern covariance with scaled dimensions (GpGp package)}
\usage{
cov_matern_scaledim(locs, covparms)
}
\arguments{
\item{locs}{A matrix of locations}

\item{covparms}{A numerical vector with covariance parameters = (variance, range_1, ..., range_d, smoothness)}
}
\value{
A matrix with \code{n} rows and \code{n} column
}
\description{
Wrapper function for Matern covariance with scaled dimensions (GpGp package)
}
\examples{
1 + 1
}
