% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmodels.R
\name{cov_matern_simple}
\alias{cov_matern_simple}
\title{Simplifed Matern covariance function}
\usage{
cov_matern_simple(d, nu, alpha = 1, tol = .Machine$double.eps)
}
\arguments{
\item{d}{A numeric vector or matrix of distances}

\item{nu}{smootheness parameter}

\item{alpha}{scale = 1/range}

\item{tol}{A numerical tolerance for the damped sine function. At \code{.Machine$double.eps} by default}
}
\value{
A matrix with \code{n} rows and \code{n} columns, with the \code{(i, j)} entry containing the simplifed matern covariance between observations \code{locs[i, ]} and \code{locs[j, ]}
}
\description{
Simplifed Matern covariance function
}
\examples{
cov_matern_simple(d = sqrt(5), nu = 2.5, alpha = 1)
cov_matern(locs = matrix(c(1, 0), 2, 1), covparms = c(1, 1, 2.5))

besselK(1, nu = 0.5)
integrate(function(x) exp(-cosh(x)) * cosh(x/2), lower = 0, upper = 1000)
}
