% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmodels.R
\name{cov_matern_spacetime}
\alias{cov_matern_spacetime}
\title{Space-time Matern covariance function}
\usage{
cov_matern_spacetime(locs, covparms, tol = .Machine$double.eps)
}
\arguments{
\item{locs}{A matrix of locations}

\item{covparms}{A numerical vector with covariance parameters = (sigma2, spatial range, temporal range, smoothness)}

\item{tol}{A numerical tolerance for the damped sine function. At \code{.Machine$double.eps} by default}
}
\value{
A matrix with \code{n} rows and \code{n} columns, with the \code{(i, j)} entry containing the space-time matern covariance between observations \code{locs[i, ]} and \code{locs[j, ]}
}
\description{
Space-time Matern covariance function
}
\examples{
locs <- matrix(runif(6), 2, 3)

cov_matern_spacetime(locs, covparms = c(1, 0.1, 0.3, 0.5))
cov_matern_spacetime_cpp(locs[1, ], locs[2, ], covparms = c(1, 0.1, 0.3, 0.5))
}
