% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cov_matern_spacetime_cpp}
\alias{cov_matern_spacetime_cpp}
\title{Space-time Matern covariance function (C++ version)}
\usage{
cov_matern_spacetime_cpp(x1, x2, covparms)
}
\arguments{
\item{x1}{First location (vector)}

\item{x2}{Second location (vector)}

\item{covparms}{A numerical vector with covariance parameters = (sigma2, spatial range, temporal range, smoothness)}
}
\value{
The space-time Matern covariance between the locations provided
}
\description{
Space-time Matern covariance function (C++ version)
}
