% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmodels.R
\name{cov_sqexpo_deriv}
\alias{cov_sqexpo_deriv}
\title{Squared exponential covariance function for GP and its derivatives}
\usage{
cov_sqexpo_deriv(locs, covparms, tol = .Machine$double.eps)
}
\arguments{
\item{locs}{A matrix of \code{n} locations}

\item{covparms}{A numerical vector with covariance parameters = c(sigma2, range).}

\item{tol}{A numerical tolerance at \code{.Machine$double.eps} by default}
}
\value{
A matrix with \code{n*d} rows and \code{n*d} columns
}
\description{
If d = 1, this function returns a squared exponential covariance matrix of (GP, dGP/dx) in 1-dimensional domain. If d = 2, this function returns a squared exponential covariance matrix of (GP, dGP/dx1, dGP/dx2) in 2-dimensional domain.
}
\examples{
locs <- matrix(runif(4), 4, 1)
covparms <- c(1, 0.1)
tol <- .Machine$double.eps

M <- cov_sqexpo_deriv(locs, covparms, tol)

t(M) == M # M is symmetric
eigen(M)$values # M is positive definite

locs <- matrix(runif(8), 4, 2)
covparms <- c(1, 0.1)
tol <- .Machine$double.eps

M <- cov_sqexpo_deriv(locs, covparms, tol)

t(M) == M # M is symmetric
eigen(M)$values # M is positive definite
}
