% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmodels.R
\name{cov_squared_expo}
\alias{cov_squared_expo}
\title{Squared exponential covariance function}
\usage{
cov_squared_expo(locs, covparms)
}
\arguments{
\item{locs}{A matrix of locations}

\item{covparms}{A numerical vector with covariance parameters}
}
\value{
A matrix with \code{n} rows and \code{n} columns, with the \code{(i, j)} entry containing the squared exponential covariance between observations \code{locs[i, ]} and \code{locs[j, ]}
}
\description{
Squared exponential covariance function
}
\examples{
locs <- matrix(runif(10), 5, 2)
all.equal(cov_squared_expo(locs = locs, covparms = c(1, 1)), 
          dnorm(fields::rdist(locs)) * sqrt(2 * pi))
}
