% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmodels.R
\name{cov_wave}
\alias{cov_wave}
\title{Wave covariance function}
\usage{
cov_wave(locs, covparms, method = "Dampedsine", tol = 1e-08)
}
\arguments{
\item{locs}{A numerical matrix with \code{n} rows and \code{p} columns. Each row of locs gives a point of the first set in R^d}

\item{covparms}{A numerical vector with covariance parameters. Its form relies on the next argument method. If you use the damped sine function then covparms = (\eqn{\sigma^2}, period). If you use the damped cosine function then covparms = (\eqn{\sigma^2}, range, period). If you use the Bessel J function then covparms = (\eqn{\sigma^2}, nu, period)}

\item{method}{A method (function) to construct a wave model. It can be "Dampedsine," "Dampedcosine," or "BesselJ." At "Dampedsine" by default}

\item{tol}{A numerical tolerance for the damped sine function. At \code{1e-8} by default}
}
\value{
A matrix with \code{n} rows and \code{n} columns, with the \code{(i, j)} entry containing the wave covariance between observations \code{locs[i, ]} and \code{locs[j, ]}
}
\description{
Wave covariance function
}
\examples{
n     <- 15^2
locs  <- matrix(runif(n * 2, 0, 1), n, 2)

h     <- fields::rdist(locs)
ind   <- order(h)
hord  <- h[ind]

covparms  <- c(1, 1/10) # covparms = c(sigma2, period) for the damped sine function
covh      <- cov_wave(locs = locs, covparms = covparms, method = "Dampedsine")
covhord   <- covh[ind]
plot(hord, covhord, type = 'l', ylim = c(-0.5, 1),
     col = 1, lwd = 2, xlab = "distance", ylab = "covariance")

covparms  <- c(1, 1/10, 1/10) # covparms = c(sigma2, range, period) for the damped cosine function
covh      <- cov_wave(locs = locs, covparms = covparms, method = "Dampedcosine")
covhord   <- covh[ind]
lines(hord, covhord, type = 'l', ylim = c(-0.5, 1), col = 2, lwd = 2)

covparms  <- c(1, 1/10, 1/10) # covparms = c(sigma2, nu, period) for the Bessel J function
covh      <- cov_wave(locs = locs, covparms = covparms, method = "BesselJ")
covhord   <- covh[ind]
lines(hord, covhord, type = 'l', ylim = c(-0.5, 1), col = 3, lwd = 2)

abline(v = 0, h = 0, col = 'gray', lwd = 2)
legend("topright", legend = c("Damped Sine", "Damped Cosine", "Bessel J"),
       col = 1:3, lty = 1, lwd = 2)
}
\references{
Abrahamsen, Petter. "A review of Gaussian random fields and correlation functions." (1997).
}
