% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia.R
\name{cvecchia_rho_specify}
\alias{cvecchia_rho_specify}
\title{specify a general vecchia approximation using the Schaefer's algorithm with known parameters}
\usage{
cvecchia_rho_specify(
  locs,
  rho,
  initial.pt = NULL,
  covmodel,
  covparms,
  abs.corr = FALSE,
  coordinate = NULL,
  ordering = "MM",
  ordering.method = "cor",
  conditioning = "NN",
  conditioning.method = "cor"
)
}
\arguments{
\item{locs}{A matrix of locations}

\item{rho}{Hyperparameter rho (instead of m)}

\item{initial.pt}{NULL = euclidean-based center, integer = specify the first obs}

\item{covmodel}{Covariance model}

\item{covparms}{Covariance parameters as a vector. The first entry must be its overall variance}

\item{abs.corr}{Logical. If \code{TRUE} then distance = 1-|rho|. If \code{FALSE} then distane = 1-rho}

\item{coordinate}{integer or vector of integers in \code{1,...,d}. This argument is used only when ordering is coord}

\item{ordering}{'MM' or 'coord'}

\item{ordering.method}{'euc' or 'cor'}

\item{conditioning}{'NN' (nearest neighbor)}

\item{conditioning.method}{'euc' or 'cor'}
}
\value{
An object that specifies the vecchia approximation for later use in likelihood evaluation or prediction
}
\description{
specify a general vecchia approximation using the Schaefer's algorithm with known parameters
}
\examples{
n             <- 20^2
m             <- 20
locs          <- matrix(runif(n * 2, 0, 1), n, 2)
covparms      <- c(1, 0.1, 10)

# true cov matrix
covmat <- cov_expo_aniso(locs, covparms)

# Visualize the process
y <- as.numeric(t(chol(covmat)) \%*\% rnorm(n))
fields::quilt.plot(locs[,1], locs[,2], y)

out01 <- cvecchia_rho_specify(locs = locs, rho = 2.5, initial.pt = NULL,
                              covmodel = cov_expo_aniso, covparms = covparms,
                              abs.corr = FALSE, coordinate = NULL,
                              ordering = "MM", ordering.method = "euc",
                              conditioning = "NN", conditioning.method = "euc")

out02 <- cvecchia_rho_specify(locs = locs, rho = 2.5, initial.pt = NULL,
                              covmodel = cov_expo_aniso, covparms = covparms,
                              abs.corr = FALSE, coordinate = NULL,
                              ordering = "MM", ordering.method = "cor",
                              conditioning = "NN", conditioning.method = "cor")

kls.euc.euc    <- performance(vecchia.approx = out01, locs = locs,
                              covmodel = cov_expo_aniso, covparms = covparms)
kls.cor.cor    <- performance(vecchia.approx = out02, locs = locs,
                              covmodel = cov_expo_aniso, covparms = covparms)

barplot(log10(c(kls.euc.euc,kls.cor.cor)),
        names.arg = c("E-MM + E-NN", "C-MM + C-NN"),
        main = "Vecchia Approximations", ylab = "log10-scale KL divergence")
}
