% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia_noneuc.R
\name{cvecchia_specify_noneuc}
\alias{cvecchia_specify_noneuc}
\title{Specify Correlation-based Vecchia approximation without locations}
\usage{
cvecchia_specify_noneuc(cormat, initial.pt = NULL, m)
}
\arguments{
\item{cormat}{Correlation matrix}

\item{initial.pt}{NULL = which.min(rowMeans(d)), integer = specify the first obs, 'random' = at random, and else = which.min(rowMeans(d))}

\item{m}{Number of nearby points to condition on}
}
\value{
An object that specifies the vecchia approximation for later use in likelihood evaluation or prediction.
}
\description{
Specify Correlation-based Vecchia approximation without locations
}
\examples{
1 + 1
}
