% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia_cpp.R
\name{fastcorrvecchia_specify_knownCovparms}
\alias{fastcorrvecchia_specify_knownCovparms}
\title{specify a general vecchia approximation using the Schaefer's algorithm with known parameters}
\usage{
fastcorrvecchia_specify_knownCovparms(
  locs,
  rho = 1.5,
  initial.pt = NULL,
  covmodel = "cov_expo_iso",
  covparms = c(1, 0.1)
)
}
\arguments{
\item{locs}{locs}

\item{rho}{rho}

\item{initial.pt}{initial.pt}

\item{covmodel}{covmodel}

\item{covparms}{covparms}
}
\value{

}
\description{
specify a general vecchia approximation using the Schaefer's algorithm with known parameters
}
\examples{
1+1
}
