% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia_noneuc.R
\name{find_ordered_cnn_noneuc}
\alias{find_ordered_cnn_noneuc}
\title{Constructing conditioning sets using only a given ordered correlation matrix}
\usage{
find_ordered_cnn_noneuc(corord, m)
}
\arguments{
\item{corord}{Correlation matrix of ordered inputs}

\item{m}{Number of nearby points to condition on}
}
\value{
Check \code{conditioning_m_Rcpp()}.
}
\description{
Constructing conditioning sets using only a given ordered correlation matrix
}
\examples{
1 + 1
}
