% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{find_ordered_nn_pred}
\alias{find_ordered_nn_pred}
\title{Constructing conditioning sets for observed and unobserved locations}
\usage{
find_ordered_nn_pred(locs, m, fix.first = 0, searchmult = 2)
}
\arguments{
\item{locs}{location matrix}

\item{m}{conditioning set size}

\item{fix.first}{fix.first}

\item{searchmult}{searchmult}
}
\value{
NNarray
}
\description{
Constructing conditioning sets for observed and unobserved locations
}
\examples{
1 + 1
}
