% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia_noneuc.R
\name{find_ordered_rn}
\alias{find_ordered_rn}
\title{Constructing conditioning sets randomly}
\usage{
find_ordered_rn(n, m)
}
\arguments{
\item{n}{Number of inputs}

\item{m}{Number of nearby points to condition on}
}
\value{
Check \code{conditioning_m_Rcpp()}
}
\description{
Constructing conditioning sets randomly
}
\examples{
1 + 1
}
