% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{fit_corrvecchia}
\alias{fit_corrvecchia}
\title{Fitting parameters using correlation-based Vecchia approximation}
\usage{
fit_corrvecchia(
  y,
  inputs,
  ms = c(30),
  trend = "zero",
  X,
  scale = "parms",
  covfun,
  var.ini,
  ranges.ini,
  nu = 3.5,
  nug = 0,
  select = Inf,
  print.level = 0,
  max.it = 32,
  tol.dec = 4,
  n.est = min(5000, nrow(inputs))
)
}
\arguments{
\item{y}{see fit_scaled}

\item{inputs}{see fit_scaled()}

\item{ms}{see fit_scaled()}

\item{trend}{see fit_scaled()}

\item{X}{see fit_scaled()}

\item{scale}{see fit_scaled()}

\item{covfun}{see fit_scaled()}

\item{var.ini}{see fit_scaled()}

\item{ranges.ini}{see fit_scaled()}

\item{nu}{see fit_scaled()}

\item{nug}{see fit_scaled()}

\item{select}{see fit_scaled()}

\item{print.level}{see fit_scaled()}

\item{max.it}{see fit_scaled()}

\item{tol.dec}{see fit_scaled()}

\item{n.est}{see fit_scaled()}
}
\value{
see fit_scaled()
}
\description{
Fitting parameters using correlation-based Vecchia approximation
}
\examples{
## generate some toy data
inputs=matrix(runif(200),ncol=2)
y=sin(rowSums(inputs*5))

## estimate parameters
fit=fit_scaled(y,inputs, trend = 'intercept')
fit$covparms

fitc <- fit_corrvecchia(y, inputs, trend = 'intercept', covfun = "matern35_scaledim")
fitc$covparms

## generate spacetime data
inputs=matrix(runif(300),ncol=3)
y=sin(rowSums(inputs*5))

fitc <- fit_corrvecchia(y, inputs, trend = 'intercept', covfun = "matern_spacetime")
fitc$covparms
}
