% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{fit_scaled}
\alias{fit_scaled}
\title{Fitting parameters using scaled Vecchia, assuming matern covariance (just copied from scaledVecchia)}
\usage{
fit_scaled(
  y,
  inputs,
  ms = c(30),
  trend = "zero",
  X,
  nu = 3.5,
  nug = 0,
  scale = "parms",
  var.ini,
  ranges.ini,
  select = Inf,
  print.level = 0,
  max.it = 32,
  tol.dec = 4,
  n.est = min(5000, nrow(inputs))
)
}
\arguments{
\item{y}{data vector of length n}

\item{inputs}{nxd matrix of input coordinates}

\item{ms}{vector of conditioning-set sizes}

\item{trend}{options are 'zero' (no trend), 'intercept', 'linear' (incl intercept)}

\item{X}{nxp trend matrix (use if more complicated trend is desired)}

\item{nu}{smoothness parameter. 1.5,2.5,3.5,4.5 avoid bessel (faster). estimated if nu=NULL}

\item{nug}{nugget or noise variance. estimated if nug=NULL}

\item{scale}{scaling of inputs for ordering and conditioning. 'parms': by parameter estimates. 'ranges': to \link{0,1}. 'none': no scaling}

\item{var.ini}{initial value for GP variance parameter}

\item{ranges.ini}{initial values for d range parameters}

\item{select}{un-select input variables if estimated range parameter is above select (assuming standardized \link{0,1} inputs)}

\item{print.level}{0: no printing. 1: print outer loop. 2: print outer+inner loop}

\item{max.it}{maximum number of iterations for inner loop}

\item{tol.dec}{converged if dot product between the step and the gradient is less than \code{10^(-convtol)}}

\item{n.est}{subsample size for estimation}
}
\value{
Object containing fit information, including for use in predictions_scaled()
}
\description{
Fitting parameters using scaled Vecchia, assuming matern covariance (just copied from scaledVecchia)
}
\examples{
inputs=matrix(runif(40),ncol=2)
y=sin(rowSums(inputs*5))
fit=fit_scaled(y,inputs)
summary.GpGp_fit(fit)
}
