% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{fit_scaled_spacetime_range}
\alias{fit_scaled_spacetime_range}
\title{Fitting range parameters using scaled Vecchia, assuming spacetime matern covariance (based on scaledVecchia)}
\usage{
fit_scaled_spacetime_range(
  method,
  y,
  inputs,
  ms = NULL,
  sig2 = 1,
  srange.ini,
  trange.ini,
  nu = 0.5,
  nug = 0,
  print.level = 0,
  max.it = 100,
  tol.dec = 4
)
}
\arguments{
\item{method}{b1, b2, b3, cc, or gp}

\item{y}{data vector of length n}

\item{inputs}{nxd matrix of input coordinates}

\item{ms}{vector of conditioning-set sizes}

\item{sig2}{GP variance parameter}

\item{srange.ini}{initial value for spatial range parameter}

\item{trange.ini}{initial value for temporal range parameter}

\item{nu}{smoothness parameter. 1.5,2.5,3.5,4.5 avoid bessel (faster). estimated if nu=NULL}

\item{nug}{nugget or noise variance. estimated if nug=NULL}

\item{print.level}{0: no printing. 1: print outer loop. 2: print outer+inner loop}

\item{max.it}{maximum number of iterations for inner loop}

\item{tol.dec}{converged if dot product between the step and the gradient is less than \code{10^(-convtol)}}
}
\value{
Object containing fit information, including for use in predictions_scaled()
}
\description{
Fitting range parameters using scaled Vecchia, assuming spacetime matern covariance (based on scaledVecchia)
}
\examples{
1 + 1
}
