% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{fit_vcf}
\alias{fit_vcf}
\title{Line search for variance correction factor}
\usage{
fit_vcf(
  fit,
  m.pred = 140,
  n.test = min(1000, round(nrow(fit$locs)/5)),
  scale = "parms",
  scorefun = lscore
)
}
\arguments{
\item{fit}{fit}

\item{m.pred}{m.pred}

\item{n.test}{n.test}

\item{scale}{scale}

\item{scorefun}{scorefun}
}
\value{
output of optimize()
}
\description{
Line search for variance correction factor
}
\examples{
1 + 1
}
