% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fun_Rcpp_example}
\alias{fun_Rcpp_example}
\title{Example function}
\usage{
fun_Rcpp_example(a, x)
}
\arguments{
\item{a}{First argument: real number}

\item{x}{Second argument: A vector}
}
\value{
Product of a and x
}
\description{
Example function
}
