% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{generate_gp}
\alias{generate_gp}
\title{Generating realizations of mean-zero Gaussian Process (GP) using a user-defined covariance function}
\usage{
generate_gp(
  nsim,
  n,
  d,
  domain,
  p = NULL,
  t = NULL,
  covmodel,
  method.locs,
  method.modify = NULL,
  pivot = FALSE,
  tol = .Machine$double.eps,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{nsim}{A number of realizations of the GP}

\item{n}{A number of locations}

\item{d}{Dimension of domain}

\item{domain}{Type of domain. It must be either "space" or "spacetime."}

\item{p}{A number of variates (GPs). At \code{NULL} by default}

\item{t}{A number of repeated measurement (= a number of different temporal locations). At \code{NULL} by default}

\item{covmodel}{A covariance function}

\item{method.locs}{If domain = space, then it must be random, overlap, or grid. If domain = spacetime, then it must be random, monitoring, grid, or satellite.}

\item{method.modify}{An argument specifying a correction method for the cholesky factorization of a covariance matrix. At \code{NULL} by default.
If correction is \code{NULL}, then the built-in R function \code{chol} is used.
If correction is \code{"qr"}, then the built-in R function \code{qr} is used.
If correction is \code{"diag"}, then \code{C + diag(tol, n)} is used instead of a covariance matrix \code{C}.
Correction methods \code{"type-I"}, \code{"type-II"}, \code{"eigen-I"}, \code{"eigen-II"}, \code{"GMW81"}, and \code{"SE99"} are refered to Fang and O'leary (2008).
Correction method \code{"nearPD"} use a built-in function nearPD() in the R package Matrix.}

\item{pivot}{Logical indicating if pivoting is to be used when factorizing a covariance matrix.  At \code{FALSE} by default}

\item{tol}{Numerical tolerance. At \code{.Machine$double.eps} by default}

\item{verbose}{Logical at \code{TRUE} by default. If verbose is \code{TRUE}, then this function prints out all messages}

\item{...}{Covariance parameters}
}
\value{
List. Please check generate_gp_space() and generate_gp_spacetime().
}
\description{
Generating realizations of mean-zero Gaussian Process (GP) using a user-defined covariance function
}
\examples{
out <- generate_gp(nsim = 1, n = 100, d = 2, domain = "space",
                   p = 1, method.locs = "random",
                   covmodel = cov_expo_iso, covparms = c(1, 0.1))

plot(as.data.frame(out$sim$sim1$locs), col = 'red', lwd = 2, main = 'Spatial locations')

out <- generate_gp(nsim = 1, n = 100, d = 2, domain = "spacetime",
                   t = 1, method.locs = "random",
                   covmodel = cov_expo_spacetime, covparms = c(1, 0.1, 1.0))

plot(as.data.frame(out$sim$sim1$locs[, 1:2]), col = 'red', lwd = 2, main = "Spatial locations")
plot(out$sim$sim1$locs[, 3], main = "Measurement time against index")
}
