% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/covmodels.R
\name{GpGp_matern_spacetime}
\alias{GpGp_matern_spacetime}
\alias{GpGp_matern_spacetime_cpp}
\title{GpGp package's matern_spacetime function (C++ version) (CAUTION: different from cov_matern_spacetime)}
\usage{
GpGp_matern_spacetime_cpp(x1, x2, covparms)

GpGp_matern_spacetime(locs, covparms)
}
\arguments{
\item{x1}{First location (vector)}

\item{x2}{Second location (vector)}

\item{covparms}{A numerical vector with covariance parameters = (sigma2, spatial range, temporal range, smoothness, nugget)}

\item{locs}{A numerical matrix with \code{n} rows and \code{d} columns. Each row of locs gives a point in R^d}
}
\value{
The space-time Matern covariance between the locations provided (CAUTION: different from cov_matern_spacetime)

A matrix with \code{n} rows and \code{n} columns, with the \code{(i, j)} entry containing the space-time matern covariance between observations \code{locs[i, ]} and \code{locs[j, ]}
}
\description{
From a location matrix \code{locs} and a vector with covariance parameters \code{covparms}, this function returns an space-time matern covariance matrix.
}
\examples{
locs <- matrix(runif(6), 2, 3)

GpGp_matern_spacetime(locs, covparms = c(1, 0.1, 0.1, 0.3, 0.5))
cov_matern_spacetime(locs, covparms = c(1, 0.1, 0.3, 0.5))

### more complex case

locs <- matrix(runif(15), 5, 3)
covparms <- c(1, 0.1, 1.0, 0.9, 0.3)

covmat1 <- GpGp_matern_spacetime(locs = locs, covparms = covparms)
covmat2 <- GpGp::matern_spacetime(covparms = covparms, locs = locs)

sum((covmat1 - covmat2)^2)

covmat1[1, 2]
GpGp_matern_spacetime_cpp(locs[1, ], locs[2, ], covparms)

covmat2[1, 2]
cov_matern_scaledim_cpp(locs[1, ], locs[2, ], c(1, 0.1, 0.1, 1, 0.9))

covparms <- c(1, 0.1, 0.2, 0.3, 0.9)

cov_matern_scaledim(locs = locs, covparms = covparms)
cov_matern_scaledim_cpp(locs[1, ], locs[2, ], covparms)
GpGp::matern_scaledim(covparms = c(covparms, 0), locs = locs)
}
