% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellany.R
\name{kldiv}
\alias{kldiv}
\title{Multivariate normal KL divergence between true distribution N(mu0, covmat0) and approximate distribution N(mu1, covmat1)}
\usage{
kldiv(
  covmat0,
  covmat1,
  mu0 = rep(0, nrow(covmat0)),
  mu1 = rep(0, nrow(covmat0)),
  chol1 = NULL
)
}
\arguments{
\item{covmat0}{A covariance matrix of the true distribution}

\item{covmat1}{A covariance matrix of the approximate distribution}

\item{mu0}{A mean vector of the true distribution assigned to zero vector by default}

\item{mu1}{A mean vector of the approximate distribution assigned to zero vector by default}

\item{chol1}{if cholesky factor of covmat1 has already been computed, provide chol1 instead of covmat1}
}
\value{
KL divergence between true and approx distributions
}
\description{
Multivariate normal KL divergence between true distribution N(mu0, covmat0) and approximate distribution N(mu1, covmat1)
}
\examples{
kldiv(diag(1, 4), diag(2, 4))
}
