% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellnay.R
\name{logscore}
\alias{logscore}
\title{Simple function for calculating logarithmic prediction score}
\usage{
logscore(mu.pred, var.pred, z)
}
\arguments{
\item{mu.pred}{predicted mean vector}

\item{var.pred}{predicted variance vector}

\item{z}{observations of the process of interest}
}
\value{
A numeric value
}
\description{
Simple function for calculating logarithmic prediction score
}
\examples{
1 + 1
}
