% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia_noneuc.R
\name{lvecchia_specify_noneuc}
\alias{lvecchia_specify_noneuc}
\title{Specify Lexicographic Vecchia approximation without locations}
\usage{
lvecchia_specify_noneuc(cormat, m)
}
\arguments{
\item{cormat}{Correlation matrix}

\item{m}{Number of nearby points to condition on}
}
\value{
An object that specifies the vecchia approximation for later use in likelihood evaluation or prediction.
}
\description{
Specify Lexicographic Vecchia approximation without locations
}
\examples{
1 + 1
}
