% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realdata.R
\name{mappingRaster}
\alias{mappingRaster}
\title{Project raster (rastername) to shape (shapename) using crs (crsname)}
\usage{
mappingRaster(rastername, tpoint, shapename, crsname = NULL, states = "all")
}
\arguments{
\item{rastername}{raster}

\item{tpoint}{timepoint}

\item{shapename}{shape}

\item{crsname}{crs}

\item{states}{states of interest}
}
\value{
list
}
\description{
Project raster (rastername) to shape (shapename) using crs (crsname)
}
\examples{
1 + 1
}
