% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mat2vals}
\alias{mat2vals}
\title{mat2vals}
\usage{
mat2vals(ptrs, inds, covmat)
}
\arguments{
\item{ptrs}{Pointers for sparse matrix}

\item{inds}{Indices for sparse matrix}

\item{covmat}{A covariance matrix}
}
\value{
A vector of covariance values
}
\description{
mat2vals
}
