% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorization.R
\name{modify}
\alias{modify}
\title{Computing a modified covariane matrix with respect to a selected correction method for positive definiteness}
\usage{
modify(
  covmat,
  pivot = FALSE,
  method = NULL,
  tol = .Machine$double.eps,
  return.err = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{covmat}{covariance matrix}

\item{pivot}{Logical indicating if pivoting is to be used when factorizing a covariance matrix.  At \code{FALSE} by default}

\item{method}{An argument specifying a correction method for the cholesky factorization of a covariance matrix. At \code{NULL} by default.
If correction is \code{NULL}, then the built-in R function \code{chol} is used.
If correction is \code{"qr"}, then the built-in R function \code{qr} is used.
If correction is \code{"diag"}, then \code{C + diag(tol, n)} is used instead of a covariance matrix \code{C}.
Correction methods \code{"type-I"}, \code{"type-II"}, \code{"eigen-I"}, \code{"eigen-II"}, \code{"GMW81"}, and \code{"SE99"} are refered to Fang and O'leary (2008).
Correction method \code{"nearPD"} use a built-in function nearPD() in the R package Matrix.}

\item{tol}{Numerical tolerance. At \code{.Machine$double.eps} by default}

\item{return.err}{Logical at \code{TRUE} by default. If return.err is true, then this function returns a decomposition error of the covariance matrix in the Frobenius norm.}

\item{verbose}{Logical at \code{FALSE} by default. If verbose is true, then this function prints out all messages.}

\item{...}{only for the function nearPD()}
}
\value{
\code{factorize} returns
\itemize{
\item{\code{covmat.modified}: } Modified covariance matrix
\item{\code{method}: } An argument specifying a correction method for the cholesky factorization of a covariance matrix.
\item{\code{tol}: } Used Numerical tolerance
\item{\code{decomp.err}: } (optional) a decomposition error of the covariance matrix
}
}
\description{
Computing a modified covariane matrix with respect to a selected correction method for positive definiteness
}
\examples{
covmat <- matrix(c(1,     0.477, 0.644, 0.478, 0.651, 0.826,
                   0.477, 1,     0.516, 0.233, 0.682, 0.75,
                   0.644, 0.516, 1,     0.599, 0.581, 0.742,
                   0.478, 0.233, 0.599, 1,     0.741, 0.8,
                   0.651, 0.682, 0.581, 0.741, 1,     0.798,
                   0.826, 0.75,  0.742, 0.8,   0.798, 1),
                 nrow = 6, ncol = 6)

modify(covmat = covmat, pivot = FALSE, method = "eigen-I", tol = .Machine$double.eps, 
       return.err = TRUE, verbose = TRUE)
modify(covmat = covmat, pivot = FALSE, method = "eigen-II", tol = .Machine$double.eps, 
       return.err = TRUE, verbose = TRUE)
modify(covmat = covmat, pivot = FALSE, method = "nearPD", tol = .Machine$double.eps, 
       return.err = TRUE, verbose = TRUE)
modify(covmat = covmat, pivot = TRUE, method = "nearPD", tol = .Machine$double.eps, 
       return.err = TRUE, verbose = TRUE)
}
