% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NNcheck_both_Rcpp}
\alias{NNcheck_both_Rcpp}
\title{creating both smaller and larger sparsity patterns (only for Florian's algorithm)}
\usage{
NNcheck_both_Rcpp(I, J, P, distances, x, rho, distype, fstr, covparms)
}
\arguments{
\item{I}{I}

\item{J}{J}

\item{P}{P}

\item{distances}{distances}

\item{x}{x}

\item{rho}{rho}

\item{distype}{distype}

\item{fstr}{fstr}

\item{covparms}{covparms}
}
\value{
list of rowvecs
}
\description{
creating both smaller and larger sparsity patterns (only for Florian's algorithm)
}
