% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NNcheck_Rcpp}
\alias{NNcheck_Rcpp}
\title{NNcheck_Rcpp (only for Florian's algorithm)}
\usage{
NNcheck_Rcpp(I, J, P, distances, x, rho, distype, fstr, covparms)
}
\arguments{
\item{I}{I}

\item{J}{J}

\item{P}{P}

\item{distances}{distances}

\item{x}{x}

\item{rho}{rho}

\item{distype}{distype}

\item{fstr}{fstr}

\item{covparms}{covparms}
}
\value{
Matrix. Check it out!
}
\description{
NNcheck_Rcpp (only for Florian's algorithm)
}
