% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia.R
\name{order_coordinate}
\alias{order_coordinate}
\title{Coordinate-based Ordering of locations}
\usage{
order_coordinate(locs, coordinate)
}
\arguments{
\item{locs}{A matrix of locations}

\item{coordinate}{integer or vector of integers in \code{1,...,d}}
}
\value{
A vector of indices giving the coordinate-based ordering
}
\description{
Coordinate-based Ordering of locations
}
\examples{
locs <- matrix(runif(100 * 2), 100, 2)

identical(order_coordinate(locs, NULL), GpGp::order_coordinate(locs, NULL))
identical(order_coordinate(locs, 1), GpGp::order_coordinate(locs, 1))
identical(order_coordinate(locs, 2), GpGp::order_coordinate(locs, 2))
identical(order_coordinate(locs, c(1, 2)), GpGp::order_coordinate(locs, c(1, 2)))
}
