% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia.R
\name{order_maxmin_correlation_inverseDist}
\alias{order_maxmin_correlation_inverseDist}
\title{Reversely implemented Maximum-minimum (maxmin) ordering of locations with respect to a user-defined distance matrix}
\usage{
order_maxmin_correlation_inverseDist(locs, d.inv, initial.pt)
}
\arguments{
\item{locs}{A matrix of locations}

\item{d.inv}{A matrix of inverse of distances between locations (for instance, rho)}

\item{initial.pt}{NULL = which.min(rowMeans(d)), center = euclidean-based center, integer = specify the first obs, 'random' = at random, and else = which.min(rowMeans(d))}
}
\value{
A vector of indices giving the correlation-based maxmin ordering
}
\description{
Reversely implemented Maximum-minimum (maxmin) ordering of locations with respect to a user-defined distance matrix
}
\examples{
covparms <- c(1, 0.1)
locs <- matrix(runif(100 * 2), 100, 2)

rho <- vecchia:::.correlation(locs, cov_expo_iso, covparms, FALSE)
identical(order_maxmin_correlation_inverseDist(locs, rho, 
                       initial.pt = GPvecchia::order_maxmin_exact(locs)[1]), 
          order_maxmin_correlation_straightforward(locs, 1 - rho, 
                       initial.pt = GPvecchia::order_maxmin_exact(locs)[1]))

covparms <- c(1, 0.1, 10)
locs.trans <- cbind(locs[ ,1] * covparms[3], locs[,2]) / covparms[2]

rho.new <- vecchia:::.correlation(locs, cov_expo_aniso, covparms, FALSE)
identical(order_maxmin_euclidean(locs.trans),
          order_maxmin_correlation_inverseDist(locs, rho.new, 
                       initial.pt = order_maxmin_euclidean(locs.trans)[1]))
}
