% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia_noneuc.R
\name{order_maxmin_correlation_noneuc}
\alias{order_maxmin_correlation_noneuc}
\title{Maximum-minimum (maxmin) ordering of locations in terms of the correlation-based distance}
\usage{
order_maxmin_correlation_noneuc(cormat, initial.pt)
}
\arguments{
\item{cormat}{Correlation matrix}

\item{initial.pt}{NULL = which.min(rowMeans(d)), integer = specify the first obs, 'random' = at random, and else = which.min(rowMeans(d))}
}
\value{
A vector of indices giving the correlation-based maxmin ordering
}
\description{
Maximum-minimum (maxmin) ordering of locations in terms of the correlation-based distance
}
\examples{
1 + 1
}
