% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia.R
\name{order_maxmin_correlation_straightforward}
\alias{order_maxmin_correlation_straightforward}
\title{Straightforward maximum-minimum (maxmin) ordering of locations with respect to a user-defined distance matrix}
\usage{
order_maxmin_correlation_straightforward(locs, d, initial.pt)
}
\arguments{
\item{locs}{A matrix of locations}

\item{d}{A matrix of distances between locations (for instance, 1-rho)}

\item{initial.pt}{NULL = which.min(rowMeans(d)), center = euclidean-based center, integer = specify the first obs, 'random' = at random, and else = which.min(rowMeans(d))}
}
\value{
A vector of indices giving the maxmin ordering with respect to the user-defined distance matrix
}
\description{
Straightforward maximum-minimum (maxmin) ordering of locations with respect to a user-defined distance matrix
}
\examples{
covparms <- c(1, 0.1)
locs <- matrix(runif(100 * 2), 100, 2)

rho <- correlationVecchia:::.correlation(locs, cov_expo_iso, covparms, FALSE)
identical(order_maxmin_euclidean(locs), 
          order_maxmin_correlation_straightforward(locs, 1 - rho, 
                                                   initial.pt = order_maxmin_euclidean(locs)[1]))
}
