% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{order_maxmin_pred}
\alias{order_maxmin_pred}
\title{Maximum-minimum (maxmin) ordering of observed and unobserved locations in terms of the Euclidean distance}
\usage{
order_maxmin_pred(locs, locs_pred, refine = FALSE)
}
\arguments{
\item{locs}{observed location matrix}

\item{locs_pred}{unobserved location matrix}

\item{refine}{logical}
}
\value{
list
}
\description{
Maximum-minimum (maxmin) ordering of observed and unobserved locations in terms of the Euclidean distance
}
\examples{
1 + 1
}
