% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{order_time}
\alias{order_time}
\title{Time-based ordering of spatio-temporal locations}
\usage{
order_time(locs, coordinate = NULL)
}
\arguments{
\item{locs}{A matrix of spatio-temporal locations}

\item{coordinate}{integer or vector of integers in \code{1,...,d}. At \code{NULL} by default}
}
\value{
A vector of indices giving the time-based ordering. If two observations are measured at the same time, then they are ordered with respect to their spatial locations.
}
\description{
Time-based ordering of spatio-temporal locations
}
\examples{
# all random
locs <- matrix(runif(30), 10, 3)
locs[order_coordinate(locs, coordinate = 3), ]
locs[order_time(locs, coordinate = NULL), ] # different

# repeated measurement
locs <- matrix(runif(20), 10, 2)
locs <- cbind(locs, sample(1:4, 10, replace = TRUE))
locs[order_coordinate(locs, coordinate = 3), ]
locs[order_time(locs, coordinate = NULL), ] # different

# In detail,
locs <- as.matrix(expand.grid(seq(4), seq(4), seq(4)))[sample(4^3), ]
all.equal(do.call(order, data.frame(locs)[, 3:1]), order(locs[, 3], locs[, 2], locs[, 1])) # same
}
