% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{order_time_old}
\alias{order_time_old}
\title{Straightforward time-based ordering of spatio-temporal locations (old version)}
\usage{
order_time_old(locs, coordinate = NULL)
}
\arguments{
\item{locs}{A matrix of spatio-temporal locations}

\item{coordinate}{integer or vector of integers in \code{1,...,d}. At \code{NULL} by default}
}
\value{
A vector of indices giving the time-based ordering. If two observations are measured at the same time, then they are ordered with respect to their spatial locations.
}
\description{
Straightforward time-based ordering of spatio-temporal locations (old version)
}
\examples{
locs <- generate_gp_spacetime(nsim = 1, n = 100, d = 2, t.len = 3, 
                              method.locs = 'all.random', 
                              covmodel = cov_spacetime_expo, covparms = c(1, 0.75, 50, 25))
locs <- locs$sim$sim1$locs

all.equal(order_time(locs), order_time_old(locs)) # same

locs <- generate_gp_spacetime(nsim = 1, n = 100, d = 2, t.len = 3, 
                              method.locs = 'space.random.time.grid', 
                              covmodel = cov_spacetime_expo, covparms = c(1, 0.75, 50, 25))
locs <- locs$sim$sim1$locs

all.equal(order_time(locs), order_time_old(locs)) # same

locs <- generate_gp_spacetime(nsim = 1, n = 100, d = 2, t.len = 3, 
                              method.locs = 'all.grid', 
                              covmodel = cov_spacetime_expo, covparms = c(1, 0.75, 50, 25))
locs <- locs$sim$sim1$locs

all.equal(order_time(locs), order_time_old(locs)) # same

locs <- generate_gp_spacetime(nsim = 1, n = 100, d = 2, t.len = 3, 
                              method.locs = 'satellite', 
                              covmodel = cov_spacetime_expo, covparms = c(1, 0.75, 50, 25))
locs <- locs$sim$sim1$locs

all.equal(order_time(locs), order_time_old(locs)) # same

# However...
locs <- as.matrix(expand.grid(seq(4), seq(4)))[sample(4^2), ]
all.equal(order_time(locs), order_time_old(locs)) # same

locs <- as.matrix(expand.grid(seq(4), seq(4), seq(4)))[sample(4^3), ]
all.equal(order_time(locs), order_time_old(locs)) # different!
}
