% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{parSim_nonst_knownCovparms}
\alias{parSim_nonst_knownCovparms}
\title{Simulation on nonstationary cases}
\usage{
parSim_nonst_knownCovparms(
  cand.m,
  nsim,
  n,
  d,
  covmodel = cov_matern_ns,
  sigma,
  smoothness,
  kernel,
  abs.corr = FALSE,
  method.locs = "random",
  method.modify = NULL,
  pivot = FALSE,
  tol = .Machine$double.eps,
  ncores = NULL
)
}
\arguments{
\item{cand.m}{A numeric vector of candidates of the size of conditioning sets (m)}

\item{nsim}{A number of repeated simualtions for each case}

\item{n}{A number of locations}

\item{d}{A dimension of domain}

\item{covmodel}{A covariance function}

\item{sigma}{numeric function for spatially varying standard deviation \eqn{\sigma( loc )}}

\item{smoothness}{A numeric function for spatially varying smoothness \eqn{\nu( loc )}}

\item{kernel}{A matrix-valued function for spatially varying (local) geometric anisotropy \eqn{\Sigma( loc )}}

\item{abs.corr}{Logical. If \code{TRUE} then distance = 1-|cor|. If \code{FALSE} then distane = 1-cor. At \code{FALSE} by default}

\item{method.locs}{random or grid}

\item{method.modify}{An argument specifying a correction method for the cholesky factorization of a covariance matrix. At \code{NULL} by default.
If correction is \code{NULL}, then the built-in R function \code{chol} is used.
If correction is \code{"qr"}, then the built-in R function \code{qr} is used.
If correction is \code{"diag"}, then \code{C + diag(tol, n)} is used instead of a covariance matrix \code{C}.
Correction methods \code{"type-I"}, \code{"type-II"}, \code{"eigen-I"}, \code{"eigen-II"}, \code{"GMW81"}, and \code{"SE99"} are refered to Fang and O'leary (2008).
Correction method \code{"nearPD"} use a built-in function nearPD() in the R package Matrix.}

\item{pivot}{Logical indicating if pivoting is to be used when factorizing a covariance matrix. At \code{FALSE} by default}

\item{tol}{Numerical tolerance. At \code{.Machine$double.eps} by default}

\item{ncores}{A number of cores for parallel computing}
}
\value{
list
}
\description{
Simulation on nonstationary cases
}
\examples{
\dontrun{
  kernel <- function(loc) {

    d         <- length(loc)

    a         <- function(loc) 10
    b         <- function(loc) 10
    angle     <- function(loc) 0

    eta       <- angle(loc)
    rot.mat   <- matrix(c(cos(eta), sin(eta), -sin(eta), cos(eta)),
                        nrow = d, ncol = d, byrow = TRUE)

    range     <- c(a(loc)^(-2), b(loc)^(-2))

    return( t(rot.mat) \%*\% diag(range, nrow = d) \%*\% rot.mat )
  }

  sigma <- function(loc) determinant(kernel(loc), logarithm = FALSE)[[1]][1]^0.25

  smoothness <- function(loc) 0.2 + 1.3 * loc[1]

  out <- parSim_nonst_knownCovparms(cand.m = c(10, 20), nsim = 2, n = 10^2, d = 2,
                                    covmodel = cov_matern_ns,
                                    sigma = sigma, smoothness = smoothness, kernel = kernel)
  out$kldiv
}
}
