% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{parSim_sptm_posterior}
\alias{parSim_sptm_posterior}
\title{parSim_sptm_posterior}
\usage{
parSim_sptm_posterior(
  cand.m,
  target,
  ordfix,
  n,
  d,
  t,
  nuggets = 0,
  method.locs,
  N = 100,
  xlim = c(0.05, 0.15),
  sdlog = 0.6,
  method.modify = NULL,
  pivot = FALSE,
  tol = .Machine$double.eps,
  verbose = FALSE,
  tol.dec = 4,
  covparms,
  ncores = NULL
)
}
\arguments{
\item{cand.m}{1}

\item{target}{1}

\item{ordfix}{1}

\item{n}{1}

\item{d}{1}

\item{t}{1}

\item{nuggets}{1}

\item{method.locs}{1}

\item{N}{1}

\item{xlim}{1}

\item{sdlog}{1}

\item{method.modify}{1}

\item{pivot}{1}

\item{tol}{1}

\item{verbose}{1}

\item{tol.dec}{tol.dec}

\item{covparms}{1}

\item{ncores}{1}
}
\value{
1
}
\description{
parSim_sptm_posterior
}
\examples{
1 + 1
}
