% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{parSim_sptm_prediction}
\alias{parSim_sptm_prediction}
\title{parSim_sptm_prediction}
\usage{
parSim_sptm_prediction(
  cand.m,
  nsim,
  n,
  n.pred,
  d,
  t,
  nuggets,
  covmodel,
  covparms,
  abs.corr,
  method.locs,
  method.locs.pred,
  method.modify = NULL,
  pivot = FALSE,
  tol = .Machine$double.eps,
  ncores = NULL
)
}
\arguments{
\item{cand.m}{1}

\item{nsim}{1}

\item{n}{1}

\item{n.pred}{1}

\item{d}{1}

\item{t}{1}

\item{nuggets}{1}

\item{covmodel}{1}

\item{covparms}{1}

\item{abs.corr}{1}

\item{method.locs}{1}

\item{method.locs.pred}{1}

\item{method.modify}{1}

\item{pivot}{1}

\item{tol}{1}

\item{ncores}{1}
}
\value{
1
}
\description{
parSim_sptm_prediction
}
\examples{
1 + 1
}
