% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellnay.R
\name{performance}
\alias{performance}
\title{Evaluating performance of Vecchia approximations in terms of KL divergence}
\usage{
performance(vecchia.approx, locs, covmodel, covparms, nuggets = 0)
}
\arguments{
\item{vecchia.approx}{object returned by specifying functions}

\item{locs}{A matrix of locations}

\item{covmodel}{A covariance model}

\item{covparms}{A numeric vector of covariance parameters}

\item{nuggets}{Nuggets at 0 by default}
}
\value{
KL divergence between true and approximate covariance matrices
}
\description{
Evaluating performance of Vecchia approximations in terms of KL divergence
}
\examples{
kldiv(diag(1, 4), diag(2, 4))
}
