% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{posterior_for_cov_matern_spacetime}
\alias{posterior_for_cov_matern_spacetime}
\title{Computing the approximate posterior of the (spatial or temporal) range parameter for cov_matern_spacetime}
\usage{
posterior_for_cov_matern_spacetime(
  z,
  locs,
  covparms,
  nugget = 0,
  m = 30,
  approx = "cc",
  target = "spatialRange",
  N = 100,
  xlim = c(0.05, 0.15),
  sdlog = 0.6
)
}
\arguments{
\item{z}{process}

\item{locs}{locs}

\item{covparms}{covariance parameters}

\item{nugget}{nugget (noise)}

\item{m}{the size of conditioning sets}

\item{approx}{b1, b2, b3, or cc}

\item{target}{spatialRange or temporalRange}

\item{N}{the number of grid points}

\item{xlim}{extent of the search}

\item{sdlog}{prior standard deviation of log(range parameter)}
}
\value{
list
}
\description{
Computing the approximate posterior of the (spatial or temporal) range parameter for cov_matern_spacetime
}
\examples{
\dontrun{
  n               <- 20^2
  covparms        <- c(1, 0.1, 0.2, 0.5)
  process         <- generate_gp_spacetime(nsim = 2, n = n, d = 2, t.len = 1, method.locs = "satellite", covmodel = cov_matern_spacetime, covparms = covparms)
  locs            <- process$sim$sim1$locs
  
  z               <- process$sim$sim1$y
  fields::quilt.plot(locs[, 1], locs[, 2], z, main = "process")
  
  covmat          <- process$sim$sim1$covmat
  
  
  par(mfrow = c(1, 2))
  
  meanlog <- log(covparms[2])
  sdlog <- 0.6
  logprior <- function(alpha) dlnorm(alpha, meanlog, sdlog, log = TRUE)
  
  plot(seq(0.01, 1, 0.01), exp(logprior(seq(0.01, 1, 0.01))), type = "l", xlab = "range parameter", ylab = "density", main = "prior distribution")
  
  meanlog <- log(covparms[3])
  sdlog <- 0.6
  logprior <- function(alpha) dlnorm(alpha, meanlog, sdlog, log = TRUE)
  
  plot(seq(0.01, 1, 0.01), exp(logprior(seq(0.01, 1, 0.01))), type = "l", xlab = "range parameter", ylab = "density", main = "prior distribution")
  
  par(mfrow = c(1, 1)) ; rm(meanlog, sdlog, logprior)
  
  
  out.spatialRange <- posterior_for_cov_matern_spacetime(z = z, locs = locs, covparms = covparms, target = "spatialRange", xlim = c(0.065, 0.14))
  out.temporalRange <- posterior_for_cov_matern_spacetime(z = z, locs = locs, covparms = covparms, target = "temporalRange", xlim = c(0.05, 0.75))
  
  par(mfrow = c(1, 2))
  matplot(out.spatialRange$alpha, out.spatialRange$post.norm, type = 'l', xlim = range(out.spatialRange$alpha), lty = 1, lwd = 2, xlab = 'range', ylab = 'density', main = "posterior distribution for spatial range parameter")
  legend('topright', c('exact', 'approx'), col = 1:2, lty = 1, lwd = 2)
  
  matplot(out.temporalRange$alpha, out.temporalRange$post.norm, type = 'l', xlim = range(out.temporalRange$alpha), lty = 1, lwd = 2, xlab = 'range', ylab = 'density', main = "posterior distribution for temporal range parameter")
  legend('topright', c('exact', 'approx'), col = 1:2, lty = 1, lwd = 2)
  par(mfrow = c(1, 1))
}
}
