% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{posterior_multivariate}
\alias{posterior_multivariate}
\title{Computing the approximate posterior of the range parameter for the mutivariate GP}
\usage{
posterior_multivariate(
  m = 30,
  approx = "cc",
  N = 150,
  limit = c(0.088, 0.118),
  z,
  locs,
  covparms,
  nugget,
  meanlog = NULL,
  sdlog = NULL,
  logprior = NULL
)
}
\arguments{
\item{m}{m}

\item{approx}{b1, b2, b3, b4, or cc}

\item{N}{N}

\item{limit}{extent of the search}

\item{z}{observations the the GP}

\item{locs}{location matrix}

\item{covparms}{covariance parameters}

\item{nugget}{nugget (noise)}

\item{meanlog}{prior mean of log(range parameter)}

\item{sdlog}{prior standard deviation of log(range parameter)}

\item{logprior}{prior function}
}
\value{
list
}
\description{
Computing the approximate posterior of the range parameter for the mutivariate GP
}
\examples{
\dontrun{
n               <- 30^2
covparms        <- c(1, 0.1, 0.1)
process         <- generate_gp_space(nsim = 1, n = n, d = 2, p = 2, method.locs = 'random', covmodel = cov_bivariate_expo_latDim, covparms = covparms)
locs            <- rbind(process$sim$sim1$locs$locs1, process$sim$sim1$locs$locs2)
nugget          <- 0.00

z               <- process$sim$sim1$y + nugget * rnorm(2 * n)
par(mfrow = c(1, 2))
fields::quilt.plot(locs[, 1][1:n], locs[, 2][1:n], z[1:n], main = "process 1")
fields::quilt.plot(locs[, 1][1:n + n], locs[, 2][1:n + n], z[1:n + n], main = "process 2")
par(mfrow = c(1, 1))

# covmat          <- process$sim$sim1$covmat + nugget * diag(2 * n)

out1 <- posterior_multivariate(m = 10, N = 150, z = z, locs = process$sim$sim1$locs, covparms = covparms, nugget = nugget)
out2 <- posterior_multivariate(m = 30, N = 150, z = z, locs = process$sim$sim1$locs, covparms = covparms, nugget = nugget)

vis_posterior_multivariate(out1 = out1, out2 = out2)
}
}
