% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{posterior_ordfix_spacetime}
\alias{posterior_ordfix_spacetime}
\title{Computing the approximate posterior of the spatial/temporal range parameter for cov_matern_spacetime}
\usage{
posterior_ordfix_spacetime(
  target = "srange",
  approx = "cc",
  method.U = "simple",
  z,
  locs,
  covparms,
  nugget = 0,
  m = 30,
  N = 100,
  xlim = c(0.05, 0.15),
  sdlog = 0.6,
  verbose = TRUE,
  tol.dec = 4
)
}
\arguments{
\item{target}{srange or trange}

\item{approx}{b1, b2, b3, cc, ex, or tr}

\item{method.U}{"simple" or "ic0"}

\item{z}{process}

\item{locs}{locs}

\item{covparms}{covariance parameters}

\item{nugget}{nugget (noise)}

\item{m}{the size of conditioning sets}

\item{N}{the number of grid points}

\item{xlim}{extent of the search}

\item{sdlog}{prior standard deviation of log(range parameter)}

\item{verbose}{logical}

\item{tol.dec}{tol.dec}
}
\value{
list
}
\description{
Computing the approximate posterior of the spatial/temporal range parameter for cov_matern_spacetime
}
\examples{
1 + 1
}
