% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction_baseline_for_spacetime}
\alias{prediction_baseline_for_spacetime}
\title{Prediction function of baseline approximations (for cov_matern_spacetime)}
\usage{
prediction_baseline_for_spacetime(
  z,
  locs,
  locs.pred,
  m,
  coordinate = NULL,
  covparms,
  nuggets,
  method = "b1",
  predcond.method = "general",
  var.exact,
  return.values = "all"
)
}
\arguments{
\item{z}{input}

\item{locs}{input}

\item{locs.pred}{input}

\item{m}{input}

\item{coordinate}{input}

\item{covparms}{input}

\item{nuggets}{input}

\item{method}{input}

\item{predcond.method}{input}

\item{var.exact}{input}

\item{return.values}{input}
}
\value{
output
}
\description{
Prediction function of baseline approximations (for cov_matern_spacetime)
}
\examples{
1 + 1
}
