% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction_corrvecchia_scaledim}
\alias{prediction_corrvecchia_scaledim}
\title{Prediction function of Vecchia approximations (for cov_matern_scaledim)}
\usage{
prediction_corrvecchia_scaledim(
  z,
  locs,
  locs.pred,
  m,
  initial.pt = NULL,
  covparms,
  nuggets,
  method = "correlation",
  predcond.method = "general",
  var.exact,
  return.values = "all"
)
}
\arguments{
\item{z}{input}

\item{locs}{input}

\item{locs.pred}{input}

\item{m}{input}

\item{initial.pt}{input}

\item{covparms}{input}

\item{nuggets}{input}

\item{method}{input}

\item{predcond.method}{input}

\item{var.exact}{input}

\item{return.values}{input}
}
\value{
output
}
\description{
Prediction function of Vecchia approximations (for cov_matern_scaledim)
}
\examples{
1 + 1
}
