% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction_spacetime_cvecchia}
\alias{prediction_spacetime_cvecchia}
\title{Prediction of Vecchia approximations for spacetime GPs}
\usage{
prediction_spacetime_cvecchia(
  method,
  z,
  locs,
  locs.pred,
  m,
  initial.pt = NULL,
  covmodel,
  covparms,
  nuggets,
  predcond.method = "general",
  var.exact,
  return.values = "all"
)
}
\arguments{
\item{method}{euc or cor}

\item{z}{A realization of the process of interest}

\item{locs}{A numeric matrix of observed locations}

\item{locs.pred}{A numeric matrix of unobserved locations}

\item{m}{The size of conditioning sets}

\item{initial.pt}{initial.pt}

\item{covmodel}{Covariance model}

\item{covparms}{A numeric vector of covariance parameters}

\item{nuggets}{nuggets}

\item{predcond.method}{It must be general for now}

\item{var.exact}{should prediction variances be computed exactly, or is a (faster) approximation acceptable}

\item{return.values}{either 'mean' only, 'meanvar', 'meanmat', or 'all'}
}
\value{
list of vecchia.approx and vecchia.predict
}
\description{
Prediction of Vecchia approximations for spacetime GPs
}
\examples{
1 + 1
}
