% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realdata.R
\name{predictions_cv_mulv}
\alias{predictions_cv_mulv}
\title{prediction using cvecchia, using output from fit_scaled_cv_mulv()}
\usage{
predictions_cv_mulv(
  ns_obs,
  ns_pred,
  fit,
  locs_pred,
  m = 100,
  joint = TRUE,
  nsims = 0,
  predvar = FALSE,
  X_pred,
  scale = "parms",
  tol = 1e-08
)
}
\arguments{
\item{ns_obs}{ns_obs}

\item{ns_pred}{ns_pred}

\item{fit}{object returned from fit_scaled_bs_mulv()}

\item{locs_pred}{ns_pred x d matrix of test/prediction inputs/locations}

\item{m}{conditioning-set size (larger is more accurate but slower)}

\item{joint}{logical}

\item{nsims}{number of simulations}

\item{predvar}{logical}

\item{X_pred}{X_pred}

\item{scale}{scaling of inputs for ordering and conditioning. 'parms': by parameter estimates. 'ranges': to (0, 1). 'none': no scaling}

\item{tol}{numerical tolerance}
}
\value{
Vector of length n.p (\code{n.sims=0}, \code{predvar=FALSE}) or list with entries \code{means} and/or \code{vars} and/or \code{samples}
}
\description{
prediction using cvecchia, using output from fit_scaled_cv_mulv()
}
\examples{
1 + 1
}
