% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predictions_scaled}
\alias{predictions_scaled}
\title{prediction using scaled Vecchia, using output from fit_scaled()}
\usage{
predictions_scaled(
  fit,
  locs_pred,
  m = 100,
  joint = TRUE,
  nsims = 0,
  predvar = FALSE,
  X_pred,
  scale = "parms",
  tol = 1e-08
)
}
\arguments{
\item{fit}{object returned from fit_scaled()}

\item{locs_pred}{n.p x d matrix of test/prediction inputs/locations}

\item{m}{conditioning-set size (larger is more accurate but slower)}

\item{joint}{Joint predictions (does not return variances) or separate/independent predictions (does not produce joint samples)}

\item{nsims}{desired number of samples from predictive distributions. if \code{n.sims=0}, posterior mean is returned.}

\item{predvar}{return prediction variances? (only if \code{joint=FALSE})}

\item{X_pred}{n.p x p trend matrix at locs_pred (if missing, will be generated based on fit object)}

\item{scale}{scaling of inputs for ordering and conditioning. 'parms': by parameter estimates. 'ranges': to the unit interval. 'none': no scaling}

\item{tol}{numerical tolerence}
}
\value{
Vector of length n.p (\code{n.sims=0}, \code{predvar=FALSE}) or list with entries \code{means} and/or \code{vars} and/or \code{samples}
}
\description{
prediction using scaled Vecchia, using output from fit_scaled()
}
\examples{
library(GpGp)

inputs=matrix(runif(200),ncol=2)
y=sin(rowSums(inputs*5))
inputs.test=matrix(runif(100),ncol=2)
fit=fit_scaled(y,inputs)
preds=predictions_scaled(fit,inputs.test)
plot(rowSums(inputs.test),preds)
}
