% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predSortSparse_Rcpp}
\alias{predSortSparse_Rcpp}
\title{Florian's Algorithm with unmeasured inputs (C++ version)}
\usage{
predSortSparse_Rcpp(xTrain, xTest, rho, initInd, distype, fstr, covparms)
}
\arguments{
\item{xTrain}{A numerical matrix of inputs (measured)}

\item{xTest}{A numerical matrix of inputs (not measured)}

\item{rho}{hyperparameter rho (instead of m)}

\item{initInd}{initial input number}

\item{distype}{euclidean or correlation}

\item{fstr}{covariance function name}

\item{covparms}{A numerical vector with covariance parameters}
}
\value{
List. Check it out!
}
\description{
Florian's Algorithm with unmeasured inputs (C++ version)
}
