% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realdata.R
\name{rasterCleaning}
\alias{rasterCleaning}
\title{Cleaning raster file}
\usage{
rasterCleaning(rasterfile)
}
\arguments{
\item{rasterfile}{raster}
}
\value{
list
}
\description{
Cleaning raster file
}
\examples{
1 + 1
}
