% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia_cpp.R
\name{rho_based_sortSparse_reverse}
\alias{rho_based_sortSparse_reverse}
\title{Schaefer's algorithm with reverse-maxmin ordering implemented in C++}
\usage{
rho_based_sortSparse_reverse(
  locs,
  rho,
  initial.pt = 1,
  covmodel = "cov_cpp",
  covparms = c(1, 0.1, 10)
)
}
\arguments{
\item{locs}{location}

\item{rho}{rho}

\item{initial.pt}{initial index}

\item{covmodel}{covmodel}

\item{covparms}{covariance parameters}
}
\value{

}
\description{
Schaefer's algorithm with reverse-maxmin ordering implemented in C++
}
\examples{
locs <- matrix(c(0.630680164, 0.785343212, 0.725967403, 0.384746804, 0.066332777,
                 0.133667622, 0.684830531, 0.642028648, 0.708973463, 0.156014202,
                 0.859438927, 0.369073307, 0.030589286, 0.249472289, 0.766846143,
                 0.762733708, 0.020763421, 0.676813798, 0.235655257, 0.278117884,
                 0.596838115, 0.61879513, 0.840685027, 0.488689207, 0.030872465,
                 0.130888217, 0.124606567, 0.050117557, 0.138916279, 0.615490804,
                 0.91239613, 0.313120618, 0.050610625, 0.990313355, 0.283005201,
                 0.029293502, 0.970113917, 0.837657897, 0.170631242, 0.038255007,
                 0.511757776, 0.174439784, 0.565152799, 0.355351824, 0.204757612,
                 0.513890368, 0.517269295, 0.791975927, 0.387010792, 0.776910901,
                 0.398383379, 0.516435312, 0.556239369, 0.418524799, 0.46378566,
                 0.681082833, 0.29392978, 0.067728284, 0.254978952, 0.337992825),
               30, 2)
rho <- 1.5
initial.pt <- 1

out <- rho_based_sortSparse_reverse(locs, rho, covmodel = "cov_expo_iso", 
                                    covparms = c(1, 0.1))
out$P
corrvecchia_specify_knownCovparms(locs, m = 5, initial.pt = 1, 
                                  covmodel = cov_expo_iso, 
                                  covparms = c(1, 0.1))$ord
}
