% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia_noneuc.R
\name{rvecchia_specify_noneuc}
\alias{rvecchia_specify_noneuc}
\title{Specify Random Vecchia approximation without locations}
\usage{
rvecchia_specify_noneuc(cormat, m)
}
\arguments{
\item{cormat}{Correlation matrix}

\item{m}{Number of nearby points to condition on}
}
\value{
An object that specifies the vecchia approximation for later use in likelihood evaluation or prediction.
}
\description{
Specify Random Vecchia approximation without locations
}
\examples{
1 + 1
}
