% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia_noneuc.R
\name{rvecchias_specify_noneuc}
\alias{rvecchias_specify_noneuc}
\title{Specify Random Vecchia approximations with different values of m but without locations}
\usage{
rvecchias_specify_noneuc(cormat, ms)
}
\arguments{
\item{cormat}{Correlation matrix}

\item{ms}{A numeric vector of numbers of nearby points to condition on}
}
\value{
An list of objects that specify the vecchia approximation for later use in likelihood evaluation or prediction.
}
\description{
Specify Random Vecchia approximations with different values of m but without locations
}
\examples{
1 + 1
}
