% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_spacetime_knownCovparms}
\alias{simulate_spacetime_knownCovparms}
\title{Conducting simulation on vecchia approximations for spacetime GPs with known covariance parameters}
\usage{
simulate_spacetime_knownCovparms(
  nsim,
  n,
  d,
  t.len,
  m,
  method.locs,
  abs.corr = FALSE,
  method.modify = NULL,
  pivot = FALSE,
  tol = .Machine$double.eps,
  verbose = TRUE,
  covmodel,
  ...
)
}
\arguments{
\item{nsim}{A number of realizations of the GP}

\item{n}{A number of locations}

\item{d}{A dimension of domain}

\item{t.len}{A number of repeated measurement (= a number of different temporal locations)}

\item{m}{A size of conditioning sets}

\item{method.locs}{"all.random", "space.random.time.grid", "all.grid", or "satellite"}

\item{abs.corr}{Logical at \code{FALSE} be default. If \code{TRUE} then distance = 1-|rho|. If \code{FALSE} then distane = 1-rho}

\item{method.modify}{An argument specifying a correction method for the cholesky factorization of a covariance matrix. At \code{NULL} by default.
If correction is \code{NULL}, then the built-in R function \code{chol} is used.
If correction is \code{"qr"}, then the built-in R function \code{qr} is used.
If correction is \code{"diag"}, then \code{C + diag(tol, n)} is used instead of a covariance matrix \code{C}.
Correction methods \code{"type-I"}, \code{"type-II"}, \code{"eigen-I"}, \code{"eigen-II"}, \code{"GMW81"}, and \code{"SE99"} are refered to Fang and O'leary (2008).
Correction method \code{"nearPD"} use a built-in function nearPD() in the R package Matrix.}

\item{pivot}{Logical indicating if pivoting is to be used when factorizing a covariance matrix. At \code{FALSE} by default}

\item{tol}{Numerical tolerance. At \code{.Machine$double.eps} by default}

\item{verbose}{Logical at \code{TRUE} by default. If verbose is \code{TRUE}, then this function prints out all messages}

\item{covmodel}{Covariance function}

\item{...}{Covariance parameters}
}
\value{
list
}
\description{
Conducting simulation on vecchia approximations for spacetime GPs with known covariance parameters
}
\examples{
out <- simulate_spacetime_knownCovparms(nsim = 2, n = 10^2, d = 2, 
                                        t.len = 2, m = 5, 
                                        method.locs = 'all.random',
                                        abs.corr = FALSE, 
                                        covmodel = cov_spacetime_expo, 
                                        covparms = c(1, 0.75, 50, 25))
out$kls.average
}
