% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_spacetime_prediction}
\alias{simulate_spacetime_prediction}
\title{simulate_spacetime_prediction}
\usage{
simulate_spacetime_prediction(
  nsim,
  n,
  n.pred,
  d,
  t,
  m,
  nuggets = 0,
  abs.corr = FALSE,
  method.locs,
  method.locs.pred,
  method.modify = NULL,
  pivot = FALSE,
  tol = .Machine$double.eps,
  verbose = FALSE,
  covmodel = cov_matern_spacetime,
  covparms
)
}
\arguments{
\item{nsim}{nsim}

\item{n}{n}

\item{n.pred}{n.pred}

\item{d}{d}

\item{t}{t}

\item{m}{m}

\item{nuggets}{nuggets}

\item{abs.corr}{logical}

\item{method.locs}{random, monitoring, or satellite}

\item{method.locs.pred}{random or subset}

\item{method.modify}{An argument specifying a correction method for the cholesky factorization of a covariance matrix. At \code{NULL} by default.
If correction is \code{NULL}, then the built-in R function \code{chol} is used.
If correction is \code{"qr"}, then the built-in R function \code{qr} is used.
If correction is \code{"diag"}, then \code{C + diag(tol, n)} is used instead of a covariance matrix \code{C}.
Correction methods \code{"type-I"}, \code{"type-II"}, \code{"eigen-I"}, \code{"eigen-II"}, \code{"GMW81"}, and \code{"SE99"} are refered to Fang and O'leary (2008).
Correction method \code{"nearPD"} use a built-in function nearPD() in the R package Matrix.}

\item{pivot}{Logical indicating if pivoting is to be used when factorizing a covariance matrix. At \code{FALSE} by default}

\item{tol}{Numerical tolerance. At \code{.Machine$double.eps} by default}

\item{verbose}{logical}

\item{covmodel}{covmodel}

\item{covparms}{covparms}
}
\value{
list
}
\description{
simulate_spacetime_prediction
}
\examples{
1 + 1
}
