% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_univariate_knownCovparms}
\alias{simulate_univariate_knownCovparms}
\title{Conducting simulation on vecchia approximations for univariate GPs with known covariance parameters}
\usage{
simulate_univariate_knownCovparms(
  nsim,
  n,
  d,
  m,
  method.locs,
  method.modify = NULL,
  corr.dist = "1-rho",
  pivot = FALSE,
  tol = .Machine$double.eps,
  verbose = TRUE,
  covmodel,
  ...
)
}
\arguments{
\item{nsim}{A number of realizations of the GP}

\item{n}{A number of locations}

\item{d}{A dimension of domain}

\item{m}{A size of conditioning sets}

\item{method.locs}{random or grid}

\item{method.modify}{An argument specifying a correction method for the cholesky factorization of a covariance matrix. At \code{NULL} by default.
If correction is \code{NULL}, then the built-in R function \code{chol} is used.
If correction is \code{"qr"}, then the built-in R function \code{qr} is used.
If correction is \code{"diag"}, then \code{C + diag(tol, n)} is used instead of a covariance matrix \code{C}.
Correction methods \code{"type-I"}, \code{"type-II"}, \code{"eigen-I"}, \code{"eigen-II"}, \code{"GMW81"}, and \code{"SE99"} are refered to Fang and O'leary (2008).
Correction method \code{"nearPD"} use a built-in function nearPD() in the R package Matrix.}

\item{corr.dist}{1-rho, 1-abs(rho), 1-rho^2, sqrt(1-rho), sqrt(1-abs(rho)), or sqrt(1-rho^2)}

\item{pivot}{Logical indicating if pivoting is to be used when factorizing a covariance matrix. At \code{FALSE} by default}

\item{tol}{Numerical tolerance. At \code{.Machine$double.eps} by default}

\item{verbose}{Logical at \code{TRUE} by default. If verbose is \code{TRUE}, then this function prints out all messages}

\item{covmodel}{Covariance function}

\item{...}{Covariance parameters}
}
\value{
list
}
\description{
Conducting simulation on vecchia approximations for univariate GPs with known covariance parameters
}
\examples{
out <- simulate_univariate_knownCovparms(nsim = 2, n = 10^2, d = 2, m = 10, 
                                         method.locs = 'random', corr.dist = "1-rho", 
                                         method.modify = NULL, 
                                         pivot = FALSE, tol = .Machine$double.eps, 
                                         verbose = TRUE, 
                                         covmodel = cov_expo_aniso, covparms = c(1, 0.1, 10))
out$kls.average
}
