% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realdata.R
\name{splitData}
\alias{splitData}
\title{Spliting a dataset into training and test datasets}
\usage{
splitData(df.joint, method, num.locs = 12, num.time = 31, size.blck = 3^2)
}
\arguments{
\item{df.joint}{dataset}

\item{method}{spacewise, timewise, or blockwise}

\item{num.locs}{number of locations}

\item{num.time}{number of time points}

\item{size.blck}{size of block}
}
\value{
list
}
\description{
Spliting a dataset into training and test datasets
}
\examples{
1 + 1
}
