% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vis_arrange}
\alias{vis_arrange}
\title{Visualization}
\usage{
vis_arrange(
  vdat1,
  vdat2 = NULL,
  legend,
  color,
  shape,
  alpha = 0.7,
  size.point = 4,
  size.line = 1,
  size.legend = 14,
  size.lab = 14,
  size.text = 12,
  size.margin = c(5.5, 20, 5.5, 5.5)
)
}
\arguments{
\item{vdat1}{First dataset consisting of the columns: index, variable (for example, m), approx_1, approx_2, ...}

\item{vdat2}{Second dataset consisting of the columns: index, variable (for example, a), approx_1, approx_2, ...}

\item{legend}{legend}

\item{color}{color}

\item{shape}{shape}

\item{alpha}{alpha}

\item{size.point}{size.point}

\item{size.line}{size.line}

\item{size.legend}{size.legend}

\item{size.lab}{size.lab}

\item{size.text}{size.text}

\item{size.margin}{size.margin}
}
\value{
output of grid.arrange(). Use plot(vis_arrange(...)) to plot it
}
\description{
Visualization
}
\examples{
\dontrun{
out <- parallel_simulate_anisotropic_knownCovparms(cand.m = c(10, 20), 
                                                   cand.a = c(1, 10), 
                                                   nsim = 2, n = 10^2, d = 2)

library(dplyr)

vdat1 <- out$vars \%>\% left_join(out$kldiv, by = "index") \%>\% 
                      filter(a == 10) \%>\% select(-a)
vdat2 <- out$vars \%>\% left_join(out$kldiv, by = "index") \%>\% 
                      filter(m == 10) \%>\% select(-m)

vis <- vis_arrange(vdat1 = vdat1, vdat2 = vdat2, 
                   legend = c("E-Maxmin + E-NN", "E-Maxmin + C-NN", 
                              "C-Maxmin + E-NN", "C-Maxmin + C-NN", 
                              "X-Coord + E-NN", "Y-Coord + E-NN"), 
                   color = c("#984EA3", "#4DAF4A", 
                             "#377EB8", "#E41A1C", 
                             "#FF7F00", "#FFFF33"), 
                   shape = c(18, 15, 17, 16, 8, 13))
ggplot2::ggsave("temp.pdf", vis, width = 10, height = 5)
}
}
