% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vis_arrange_posterior}
\alias{vis_arrange_posterior}
\title{Visualization}
\usage{
vis_arrange_posterior(
  output,
  legend = c("T-ord + T-NN", "T-ord + E-NN", "T-ord + C-NN", "C-MM + C-NN", "True"),
  color = c("#984EA3", "#4DAF4A", "#377EB8", "#E41A1C", "black"),
  xlim = NULL,
  ylim = NULL,
  xlab = "x",
  ylab = "y",
  alpha = 0.7,
  size.line = 1,
  size.legend = 14,
  size.lab = 14,
  size.text = 12,
  size.margin = c(5.5, 20, 5.5, 5.5)
)
}
\arguments{
\item{output}{output from posterior_... functions}

\item{legend}{legend}

\item{color}{color}

\item{xlim}{xlim}

\item{ylim}{ylim}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{alpha}{alpha}

\item{size.line}{size.line}

\item{size.legend}{size.legend}

\item{size.lab}{size.lab}

\item{size.text}{size.text}

\item{size.margin}{size.margin}
}
\value{
plot
}
\description{
Visualization
}
\examples{
1 + 1
}
