% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vis_arrange_tetra}
\alias{vis_arrange_tetra}
\title{Visualization}
\usage{
vis_arrange_tetra(
  vdat1,
  vdat2,
  vdat3,
  vdat4,
  legend,
  color,
  shape,
  ftn = log10,
  xlabs = c("m", "m", "m", "m"),
  ylabs = c("log10(KL)", "log10(KL)", "log10(KL)", "log10(KL)"),
  xlims = list(xl1 = NULL, xl2 = NULL, xl3 = NULL, xl4 = NULL),
  ylims = list(yl1 = NULL, yl2 = NULL, yl3 = NULL, yl4 = NULL),
  alpha = 0.7,
  size.point = 2,
  size.line = 1,
  size.legend = 14,
  size.lab = 14,
  size.text = 12,
  size.margin = c(5.5, 20, 5.5, 5.5)
)
}
\arguments{
\item{vdat1}{vdat1}

\item{vdat2}{vdat2}

\item{vdat3}{vdat3}

\item{vdat4}{vdat4}

\item{legend}{legend}

\item{color}{color}

\item{shape}{shape}

\item{ftn}{ftn}

\item{xlabs}{xlabs}

\item{ylabs}{ylabs}

\item{xlims}{xlims}

\item{ylims}{ylims}

\item{alpha}{alpha}

\item{size.point}{size.point}

\item{size.line}{size.line}

\item{size.legend}{size.legend}

\item{size.lab}{size.lab}

\item{size.text}{size.text}

\item{size.margin}{size.margin}
}
\value{
plot
}
\description{
Visualization
}
\examples{
1 + 1
}
