% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{vis_posterior_multivariate}
\alias{vis_posterior_multivariate}
\title{Visualizing the approximate posteriors of the range parameter for the mutivariate GP}
\usage{
vis_posterior_multivariate(
  out1,
  out2,
  color = c("#377EB8", "#E41A1C", "gray1"),
  legend = c("m = 10", "m = 30", "True density"),
  ylimit = c(0, 0.025),
  alpha = 0.7,
  size.point = 4,
  size.line = 1,
  size.legend = 14,
  size.lab = 14,
  size.text = 12,
  size.margin = c(5.5, 20, 5.5, 5.5),
  tol = 1e-08
)
}
\arguments{
\item{out1}{out1}

\item{out2}{out2}

\item{color}{color}

\item{legend}{legend}

\item{ylimit}{ylimit}

\item{alpha}{alpha}

\item{size.point}{size.point}

\item{size.line}{size.line}

\item{size.legend}{size.legend}

\item{size.lab}{size.lab}

\item{size.text}{size.text}

\item{size.margin}{size.margin}
}
\value{
ggplot
}
\description{
Visualizing the approximate posteriors of the range parameter for the mutivariate GP
}
\examples{
\dontrun{
n               <- 30^2
covparms        <- c(1, 0.1, 0.1)
process         <- generate_gp_space(nsim = 1, n = n, d = 2, p = 2, method.locs = 'random', covmodel = cov_bivariate_expo_latDim, covparms = covparms)
locs            <- rbind(process$sim$sim1$locs$locs1, process$sim$sim1$locs$locs2)
nugget          <- 0.00

z               <- process$sim$sim1$y + nugget * rnorm(2 * n)
par(mfrow = c(1, 2))
fields::quilt.plot(locs[, 1][1:n], locs[, 2][1:n], z[1:n], main = "process 1")
fields::quilt.plot(locs[, 1][1:n + n], locs[, 2][1:n + n], z[1:n + n], main = "process 2")
par(mfrow = c(1, 1))

# covmat          <- process$sim$sim1$covmat + nugget * diag(2 * n)

out1 <- posterior_multivariate(m = 10, N = 150, z = z, locs = process$sim$sim1$locs, covparms = covparms, nugget = nugget)
out2 <- posterior_multivariate(m = 30, N = 150, z = z, locs = process$sim$sim1$locs, covparms = covparms, nugget = nugget)

vis_posterior_multivariate(out1 = out1, out2 = out2)
}
}
