% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{vis_prediction}
\alias{vis_prediction}
\title{Visualization function for prediction}
\usage{
vis_prediction(
  vdat1,
  vdat2,
  vdat3,
  ylab_vdat1 = "log10( logscore )",
  ylab_vdat2 = "log10( logscore )",
  ylab_vdat3 = "log10( logscore )",
  legend,
  color,
  shape,
  alpha = 0.7,
  size.point = 4,
  size.line = 1,
  size.legend = 14,
  size.lab = 14,
  size.text = 12,
  size.margin = c(5.5, 20, 5.5, 5.5)
)
}
\arguments{
\item{vdat1}{input}

\item{vdat2}{input}

\item{vdat3}{input}

\item{ylab_vdat1}{input}

\item{ylab_vdat2}{input}

\item{ylab_vdat3}{input}

\item{legend}{input}

\item{color}{input}

\item{shape}{input}

\item{alpha}{input}

\item{size.point}{input}

\item{size.line}{input}

\item{size.legend}{input}

\item{size.lab}{input}

\item{size.text}{input}

\item{size.margin}{input}
}
\value{
output
}
\description{
Visualization function for prediction
}
\examples{
1 + 1
}
