% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia.R
\name{corrvecchia_specify_knownCovparms_2}
\alias{corrvecchia_specify_knownCovparms_2}
\title{specify Euclidean/Correlation-based Vecchia approximation with known parameters}
\usage{
corrvecchia_specify_knownCovparms_2(
  locs,
  m,
  ordering = "maxmin",
  ordering.method = "correlation",
  coordinate = NULL,
  corr.dist = "1-rho",
  initial.pt = NULL,
  conditioning = "NN",
  conditioning.method = "correlation",
  covmodel,
  covparms
)
}
\arguments{
\item{locs}{A matrix of locations}

\item{m}{Number of nearby points to condition on}

\item{ordering}{'maxmin' or 'coord'}

\item{ordering.method}{either 'euclidean' or 'correlation'}

\item{coordinate}{integer or vector of integers in \code{1,...,d}}

\item{corr.dist}{1-rho, 1-abs(rho), 1-rho^2, sqrt(1-rho), sqrt(1-abs(rho)), or sqrt(1-rho^2)}

\item{initial.pt}{NULL = which.min(rowMeans(d)), center = euclidean-based center, integer = specify the first obs, 'random' = at random, and else = which.min(rowMeans(d))}

\item{conditioning}{'NN' (nearest neighbor)}

\item{conditioning.method}{either 'euclidean' or 'correlation'}

\item{covmodel}{covariance function (or matrix)}

\item{covparms}{covariance parameters as a vector. The first entry must be its overall variance}
}
\value{
An object that specifies the vecchia approximation for later use in likelihood evaluation or prediction.
}
\description{
specify Euclidean/Correlation-based Vecchia approximation with known parameters
}
