% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmodels.R
\name{cov_bivariate_expo_latDim}
\alias{cov_bivariate_expo_latDim}
\title{Isotropic exponontial covariance function for bivariate Gaussian Processes using latent coordinate}
\usage{
cov_bivariate_expo_latDim(locs1, locs2, covparms)
}
\arguments{
\item{locs1}{A numerical matrix with \code{n1} rows and \code{d} columns. Each row of locs1 gives a point of the first set in R^d}

\item{locs2}{A numerical matrix with \code{n2} rows and \code{d} columns. Each row of locs2 gives a point of the second set in R^d}

\item{covparms}{A numerical vector with covariance parameters. Its form must be (variance, range, distance between two processes in latent coordinate).}
}
\value{
An isotropic exponential covariance matrix with \code{n1 + n2} rows and \code{n1 + n2} columns
}
\description{
Isotropic exponontial covariance function for bivariate Gaussian Processes using latent coordinate
}
\examples{
locs1 <- matrix(runif(6), 3, 2)
locs2 <- matrix(runif(8), 4, 2)
covparms <- c(1, 0.1)
d.latent <- 1

cov_bivariate_expo_latDim(locs1 = locs1, locs2 = locs2, covparms = c(covparms, d.latent))
}
