% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmodels.R
\name{cov_matern_aniso}
\alias{cov_matern_aniso}
\title{Anisotropic Matern covariance function}
\usage{
cov_matern_aniso(locs, covparms, a = NULL, tol = .Machine$double.eps)
}
\arguments{
\item{locs}{A numerical matrix with \code{n} rows and \code{d} columns. Each row of locs gives a point in R^d}

\item{covparms}{A numerical vector with covariance parameters. It can be of two different forms: One is the 4-dim'l vector (variance, range, smoothness, degree of anisotropy) and the other is simply 3-dim'l vector (variance, range, smoothness). In the latter case, a vector of anisotropy must be specified}

\item{a}{A vector of anisotropy. At \code{NULL} by default}

\item{tol}{A numerical tolerance for the damped sine function. At \code{.Machine$double.eps} by default}
}
\value{
A matrix with \code{n} rows and \code{n} columns, with the \code{(i, j)} entry containing the anisotropic matern covariance between observations \code{locs[i, ]} and \code{locs[j, ]}
}
\description{
From a location matrix \code{locs} and a vector with covariance parameters \code{covparms}, this function returns an anisotropic matern covariance matrix.
}
\examples{
locs <- matrix(runif(4), 2, 2)

cov_matern_aniso(locs, covparms = c(1, 1, 0.5, 5))
cov_matern_scaledim(locs, covparms = c(1, 1/5, 1, 0.5))
}
