% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellnay.R
\name{kldiv}
\alias{kldiv}
\title{Multivariate normal KL divergence between true distribution N(mu0, covmat0) and approximate distribution N(mu1, covmat1)}
\usage{
kldiv(
  covmat0,
  covmat1,
  mu0 = rep(0, nrow(covmat0)),
  mu1 = rep(0, nrow(covmat0)),
  chol1 = NULL,
  method.modify = NULL,
  pivot = FALSE,
  tol = .Machine$double.eps
)
}
\arguments{
\item{covmat0}{A covariance matrix of the true distribution}

\item{covmat1}{A covariance matrix of the approximate distribution}

\item{mu0}{A mean vector of the true distribution assigned to zero vector by default}

\item{mu1}{A mean vector of the approximate distribution assigned to zero vector by default}

\item{chol1}{if cholesky factor of covmat1 has already been computed, provide chol1 instead of covmat1}

\item{method.modify}{An argument specifying a correction method for the cholesky factorization of a covariance matrix. At \code{NULL} by default.
If correction is \code{NULL}, then the built-in R function \code{chol} is used.
If correction is \code{"qr"}, then the built-in R function \code{qr} is used.
If correction is \code{"diag"}, then \code{C + diag(tol, n)} is used instead of a covariance matrix \code{C}.
Correction methods \code{"type-I"}, \code{"type-II"}, \code{"eigen-I"}, \code{"eigen-II"}, \code{"GMW81"}, and \code{"SE99"} are refered to Fang and O'leary (2008).
Correction method \code{"nearPD"} use a built-in function nearPD() in the R package Matrix.}

\item{pivot}{Logical indicating if pivoting is to be used when factorizing a covariance matrix. At \code{FALSE} by default}

\item{tol}{Numerical tolerance. At \code{.Machine$double.eps} by default}
}
\value{
KL divergence between true and approx distributions
}
\description{
Multivariate normal KL divergence between true distribution N(mu0, covmat0) and approximate distribution N(mu1, covmat1)
}
\examples{
kldiv(diag(1, 4), diag(2, 4))
}
