% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellany.R
\name{performance}
\alias{performance}
\title{Evaluating performance of Vecchia approximations in terms of KL divergence}
\usage{
performance(out, locs, covmodel, covparms)
}
\arguments{
\item{out}{vecchia_specify}

\item{locs}{a matrix of locations}

\item{covmodel}{covariance function}

\item{covparms}{a numeric vector of covariance parameters}
}
\value{
KL divergence between true and approx distributions
}
\description{
Evaluating performance of Vecchia approximations in terms of KL divergence
}
